Move Workflow Session domain tests out of root lib-test
