+++ src/tool_runtime/sessions/tests.rs
@@ -1,13 +1,10 @@
 use super::super::project_instructions::{LoadedInstructionCandidate, ProjectInstructionsSnapshot};
-use super::super::tool_inputs::{ExecutionShell, SessionMode};
+use super::super::tool_inputs::SessionMode;
 use super::events::{
     changed_paths_for_tool, normalize_observed_project_path, observed_paths_for_successful_result,
     session_input_summary_for_tool, SessionToolClassification,
 };
-use super::model::{
-    PersistedSessionLedger, SessionLifecycle, MAX_OBSERVED_PATHS_PER_EVENT,
-    MAX_VALIDATION_EXCERPT_CHARS, MESSAGE_ID_PREFIX, SESSION_LEDGER_VERSION,
-};
+use super::model::{PersistedSessionLedger, MAX_OBSERVED_PATHS_PER_EVENT, SESSION_LEDGER_VERSION};
 use super::persistence::write_ledger_atomic;
 use super::*;
 use serde_json::{json, Value};
@@ -46,6 +43,72 @@ fn record_model_facing_result(
             (!success).then_some("business_failure"),
         )
         .expect("recorded model-facing result")
+}
+
+fn persistent_store(path: PathBuf) -> SessionStore {
+    SessionStore::with_persistence(path, 10, 10)
+}
+
+fn flush_and_restore(store: &SessionStore, path: PathBuf) -> SessionStore {
+    store.flush_persistence();
+    SessionStore::with_persistence(path, 10, 10)
+}
+
+fn record_console_tool(
+    store: &SessionStore,
+    session_id: &str,
+    project: &str,
+    tool: &str,
+    input: Value,
+    success: bool,
+    output: Value,
+) {
+    let start = store.record_tool_call_started(
+        Some(session_id),
+        SessionTransport::Api,
+        tool,
+        &input,
+        crate::tool_runtime::sessions::session_tool_contract(tool),
+    );
+    store.record_tool_call_finished(
+        start,
+        success,
+        &output,
+        (!success).then_some("tool failed"),
+        (!success).then_some("validation_failed"),
+    );
+    debug_assert_eq!(input.get("project").and_then(Value::as_str), Some(project));
+}
+
+fn record_console_validation(
+    store: &SessionStore,
+    session_id: &str,
+    project: &str,
+    success: bool,
+    tests_run_count: u64,
+) {
+    record_console_tool(
+        store,
+        session_id,
+        project,
+        "cargo_test",
+        json!({"project": project}),
+        success,
+        json!({
+            "exit_code": if success { 0 } else { 101 },
+            "execution_state": "completed",
+            "command_started": true,
+            "command_completed": true,
+            "tests_detected": true,
+            "tests_run_count": tests_run_count,
+            "zero_tests_run": tests_run_count == 0,
+            "stdout_tail": if success { "test result: ok" } else { "test result: FAILED" },
+            "stderr_tail": "",
+            "stdout_lines": 1,
+            "stderr_lines": 0,
+            "failure_kind": if success { Value::Null } else { json!("test_failure") }
+        }),
+    );
 }
 
 #[test]
@@ -509,82 +572,6 @@ fn exploration_input_audit_omits_queries_and_shell_commands() {
 }
 
 #[test]
-fn session_store_bounds_event_limit() {
-    let store = SessionStore::new(10, 3);
-    let summary = store.start_session(None, None);
-    for idx in 0..5 {
-        let args = json!({"project": "demo", "path": format!("file{idx}.rs")});
-        let start = store.record_tool_call_started(
-            Some(&summary.session_id),
-            SessionTransport::Api,
-            "write_project_file",
-            &args,
-            crate::tool_runtime::sessions::session_tool_contract("write_project_file"),
-        );
-        store.record_tool_call_finished(start, true, &json!({}), None, None);
-    }
-    let summary = store.summary(&summary.session_id, Some(50)).unwrap();
-    assert_eq!(summary.events.len(), 3);
-    assert_eq!(summary.counts.tool_calls, 2);
-}
-
-#[test]
-fn input_summary_redacts_sensitive_keys() {
-    let store = SessionStore::default();
-    let summary = store.start_session(None, None);
-    store.record_tool_call_started(
-        Some(&summary.session_id),
-        SessionTransport::Api,
-        "read_file",
-        &json!({
-            "project": "demo",
-            "token": "super-secret-token",
-            "command": "curl -H 'Authorization: Bearer wc_pat_never_store'"
-        }),
-        crate::tool_runtime::sessions::session_tool_contract("read_file"),
-    );
-    let summary = store.summary(&summary.session_id, Some(10)).unwrap();
-    assert_eq!(
-        summary.events[0].input_summary.as_ref().unwrap()["token"],
-        "[redacted]"
-    );
-    assert_eq!(
-        summary.events[0].input_summary.as_ref().unwrap()["command"],
-        "[redacted]"
-    );
-}
-
-#[test]
-fn coding_instruction_redacts_reusable_credentials_without_truncating_normal_goals() {
-    assert_eq!(
-        super::util::redact_and_bound_instruction(
-            "continue with wc_pat_never_persist_this_value",
-            super::model::MAX_CODING_INSTRUCTION_CHARS,
-        ),
-        "[redacted]"
-    );
-    let goal = "x".repeat(1_000);
-    assert_eq!(
-        super::util::redact_and_bound_instruction(
-            &goal,
-            super::model::MAX_CODING_INSTRUCTION_CHARS,
-        ),
-        goal
-    );
-}
-
-fn persistent_store(path: PathBuf) -> SessionStore {
-    SessionStore::with_persistence(path, 10, 10)
-}
-
-/// Flush deferred ledger writes, then open a fresh store from the same path.
-/// Required because persistent stores write on a background thread.
-fn flush_and_restore(store: &SessionStore, path: PathBuf) -> SessionStore {
-    store.flush_persistence();
-    SessionStore::with_persistence(path, 10, 10)
-}
-
-#[test]
 fn coding_agent_lifecycle_evidence_is_project_scoped_body_free_and_durable() {
     let tmp = tempfile::tempdir().unwrap();
     let ledger = tmp.path().join("sessions.json");
@@ -667,38 +654,6 @@ fn coding_agent_lifecycle_evidence_is_project_scoped_body_free_and_durable() {
     assert!(!serialized.contains("idempotency_key"));
     assert!(!serialized.contains("acp_session"));
     assert!(!serialized.contains("wc_agent_run_wrongproject"));
-}
-
-#[test]
-fn session_store_persists_and_restores_basic_session() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger = tmp.path().join("sessions.json");
-    let store = persistent_store(ledger.clone());
-    let session = store.start_session(
-        Some("agent:oe:private-drop".to_string()),
-        Some("persistent work".to_string()),
-    );
-
-    store.flush_persistence();
-    let raw = std::fs::read_to_string(&ledger).unwrap();
-    assert!(
-        !raw.contains('\n'),
-        "production ledger should use compact JSON"
-    );
-    let restored = SessionStore::with_persistence(ledger, 10, 10);
-    let status = restored.status();
-    assert_eq!(status.persistence, "enabled");
-    assert_eq!(status.restored_sessions, 1);
-    assert_eq!(status.last_persist_error, None);
-    let summary = restored.summary(&session.session_id, Some(10)).unwrap();
-    assert_eq!(summary.session_id, session.session_id);
-    assert_eq!(summary.project.as_deref(), Some("agent:oe:private-drop"));
-    assert_eq!(summary.title.as_deref(), Some("persistent work"));
-    assert_eq!(summary.lifecycle, SessionLifecycle::Active);
-    assert_eq!(
-        summary.execution_context,
-        SessionExecutionContext::default()
-    );
 }
 
 #[test]
@@ -1069,501 +1024,6 @@ fn write_ledger_atomic_cleans_up_temp_file_when_rename_fails() {
 }
 
 #[test]
-fn session_execution_context_persistence_matrix() {
-    let cases = [
-        (
-            "local-normalized",
-            SessionExecutionContext {
-                default_cwd: Some("frontend/./src".to_string()),
-                default_shell: Some(ExecutionShell::Bash),
-                resource: None,
-            },
-            SessionExecutionContext {
-                default_cwd: Some("frontend/src".to_string()),
-                default_shell: Some(ExecutionShell::Bash),
-                resource: None,
-            },
-            json!({"default_cwd": "frontend/src", "default_shell": "bash"}),
-        ),
-        (
-            "remote-resource",
-            SessionExecutionContext {
-                default_cwd: Some("/opt/webcodex-edge".to_string()),
-                default_shell: None,
-                resource: Some("tmp".to_string()),
-            },
-            SessionExecutionContext {
-                default_cwd: Some("/opt/webcodex-edge".to_string()),
-                default_shell: None,
-                resource: Some("tmp".to_string()),
-            },
-            json!({"default_cwd": "/opt/webcodex-edge", "resource": "tmp"}),
-        ),
-    ];
-
-    for (label, input, expected, persisted_context) in cases {
-        let tmp = tempfile::tempdir().unwrap();
-        let ledger = tmp.path().join("sessions.json");
-        let store = persistent_store(ledger.clone());
-        let session = store
-            .start_session_with_options(
-                SessionCreateOptions::new(
-                    Some("agent:oe:private-drop".to_string()),
-                    Some(format!("persistent context {label}")),
-                    SessionMode::Normal,
-                    SessionGuards::default(),
-                )
-                .with_execution_context(input),
-            )
-            .unwrap();
-        assert_eq!(session.execution_context, expected, "{label}");
-
-        store.flush_persistence();
-        let value: Value = serde_json::from_slice(&std::fs::read(&ledger).unwrap()).unwrap();
-        assert_eq!(
-            value["sessions"][0]["execution_context"], persisted_context,
-            "{label}"
-        );
-        let restored = SessionStore::with_persistence(ledger.clone(), 10, 10);
-        assert_eq!(
-            restored
-                .summary(&session.session_id, None)
-                .unwrap()
-                .execution_context,
-            expected,
-            "{label}"
-        );
-    }
-}
-
-#[test]
-fn coding_session_context_precommit_failure_leaves_memory_unchanged() {
-    fn request(
-        resume_session_id: Option<String>,
-        execution_context: Option<SessionExecutionContext>,
-    ) -> CodingSessionRequest {
-        CodingSessionRequest {
-            project: "agent:oe:private-drop".to_string(),
-            authority_fingerprint: TEST_ONLY_PROJECT_SESSION_AUTHORITY_FINGERPRINT.to_string(),
-            resume_session_id,
-            instruction: Some("continue".to_string()),
-            mode: SessionMode::Normal,
-            guards: SessionGuards::default(),
-            execution_context,
-            project_instructions: None,
-            transport: SessionTransport::Api,
-            context_refreshed: true,
-            write_scope_verified: true,
-        }
-    }
-
-    let store = SessionStore::default();
-    let initial = SessionExecutionContext {
-        default_cwd: Some("frontend".to_string()),
-        default_shell: Some(ExecutionShell::Bash),
-        resource: None,
-    };
-    let created = store
-        .ensure_coding_session(request(None, Some(initial.clone())))
-        .unwrap();
-    assert!(!created.reused);
-    assert!(created.execution_context_changed);
-    assert_eq!(created.summary.execution_context, initial);
-
-    let session_id = created.summary.session_id.clone();
-    let preserved = store
-        .ensure_coding_session(request(Some(session_id.clone()), None))
-        .unwrap();
-    assert!(preserved.reused);
-    assert!(!preserved.execution_context_changed);
-    assert_eq!(preserved.summary.execution_context, initial);
-
-    let replacement = SessionExecutionContext {
-        default_cwd: Some("backend".to_string()),
-        default_shell: Some(ExecutionShell::Sh),
-        resource: None,
-    };
-    let replaced = store
-        .ensure_coding_session(request(Some(session_id.clone()), Some(replacement.clone())))
-        .unwrap();
-    assert!(replaced.execution_context_changed);
-    assert_eq!(replaced.summary.execution_context, replacement);
-    let event = replaced.summary.events.last().unwrap();
-    assert_eq!(event.execution_context, Some(replacement.clone()));
-    assert_eq!(event.previous_execution_context, Some(initial.clone()));
-
-    let event_count = replaced.summary.events_total;
-    store.fail_next_coding_continuity_precommit_for_test();
-    let error = store
-        .ensure_coding_session(request(
-            Some(session_id.clone()),
-            Some(SessionExecutionContext::default()),
-        ))
-        .unwrap_err();
-    assert_eq!(error, CodingSessionError::CommitFailed);
-    let after_failure = store.summary(&session_id, None).unwrap();
-    assert_eq!(after_failure.execution_context, replacement);
-    assert_eq!(after_failure.events_total, event_count);
-}
-
-#[test]
-fn update_session_execution_context_sets_clears_and_rejects_invalid_states() {
-    let store = SessionStore::default();
-    let session = store.start_session(
-        Some("agent:oe:private-drop".to_string()),
-        Some("context update".to_string()),
-    );
-    let replacement = SessionExecutionContext {
-        default_cwd: Some("frontend".to_string()),
-        default_shell: Some(ExecutionShell::Bash),
-        resource: None,
-    };
-    let set = store
-        .update_execution_context(
-            &session.session_id,
-            replacement.clone(),
-            SessionTransport::Api,
-        )
-        .unwrap();
-    assert!(set.changed);
-    assert_eq!(
-        set.previous_execution_context,
-        SessionExecutionContext::default()
-    );
-    assert_eq!(set.summary.execution_context, replacement);
-    assert_eq!(
-        set.summary.events.last().unwrap().kind,
-        "session_execution_context_updated"
-    );
-
-    let cleared = store
-        .update_execution_context(
-            &session.session_id,
-            SessionExecutionContext::default(),
-            SessionTransport::Mcp,
-        )
-        .unwrap();
-    assert!(cleared.changed);
-    assert_eq!(
-        cleared.summary.execution_context,
-        SessionExecutionContext::default()
-    );
-
-    let before_invalid = cleared.summary.events_total;
-    let invalid = store
-        .update_execution_context(
-            &session.session_id,
-            SessionExecutionContext {
-                default_cwd: Some("../outside".to_string()),
-                default_shell: None,
-                resource: None,
-            },
-            SessionTransport::Api,
-        )
-        .unwrap_err();
-    assert!(matches!(
-        invalid,
-        SessionExecutionContextUpdateError::InvalidExecutionContext(_)
-    ));
-    assert_eq!(
-        store
-            .summary(&session.session_id, None)
-            .unwrap()
-            .events_total,
-        before_invalid
-    );
-
-    assert_eq!(
-        store
-            .update_execution_context(
-                "wc_sess_missingcontext01",
-                SessionExecutionContext::default(),
-                SessionTransport::Api,
-            )
-            .unwrap_err(),
-        SessionExecutionContextUpdateError::UnknownSession
-    );
-    let unscoped = store.start_session(None, Some("unscoped".to_string()));
-    assert_eq!(
-        store
-            .update_execution_context(
-                &unscoped.session_id,
-                SessionExecutionContext::default(),
-                SessionTransport::Api,
-            )
-            .unwrap_err(),
-        SessionExecutionContextUpdateError::SessionHasNoProject
-    );
-    store.close_session(&session.session_id).unwrap();
-    assert_eq!(
-        store
-            .update_execution_context(
-                &session.session_id,
-                SessionExecutionContext::default(),
-                SessionTransport::Api,
-            )
-            .unwrap_err(),
-        SessionExecutionContextUpdateError::SessionNotActive {
-            lifecycle: SessionLifecycle::Closed
-        }
-    );
-}
-
-#[test]
-fn persistent_shell_evidence_survives_restore_without_command_or_output() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger = tmp.path().join("sessions.json");
-    let store = persistent_store(ledger.clone());
-    let session = store.start_session(None, Some("persistent shell evidence".to_string()));
-    let start = store.record_tool_call_started(
-        Some(&session.session_id),
-        SessionTransport::Api,
-        "session_shell_exec",
-        &json!({
-            "project": "agent:oe:private-drop",
-            "session_id": session.session_id.clone(),
-            "shell_id": "wc_shell_evidence",
-            "command": "export PRIVATE_LEDGER_VALUE=secret",
-            "command_summary": "export PRIVATE_LEDGER_VALUE=secret",
-            "command_present": true
-        }),
-        crate::tool_runtime::sessions::session_tool_contract("session_shell_exec"),
-    );
-    store.record_tool_call_finished(
-        start,
-        true,
-        &json!({
-            "shell_id": "wc_shell_evidence",
-            "shell_state": "running",
-            "execution_state": "completed",
-            "command_started": true,
-            "command_completed": true,
-            "exit_code": 0,
-            "stdout": "PRIVATE_LEDGER_VALUE=secret",
-            "stderr": ""
-        }),
-        None,
-        None,
-    );
-    store.flush_persistence();
-    let persisted = std::fs::read_to_string(&ledger).unwrap();
-    assert!(!persisted.contains("PRIVATE_LEDGER_VALUE"));
-    assert!(!persisted.contains("\"stdout\""));
-    assert!(!persisted.contains("\"stderr\""));
-
-    let restored = SessionStore::with_persistence(ledger, 10, 10);
-    let summary = restored.summary(&session.session_id, Some(10)).unwrap();
-    let evidence = summary.events[1].persistent_shell.as_ref().unwrap();
-    assert_eq!(evidence.action, "exec");
-    assert_eq!(evidence.shell_id.as_deref(), Some("wc_shell_evidence"));
-    assert_eq!(evidence.shell_state.as_deref(), Some("running"));
-    assert_eq!(evidence.execution_state.as_deref(), Some("completed"));
-    assert_eq!(evidence.command_started, Some(true));
-    assert_eq!(evidence.command_completed, Some(true));
-}
-
-#[test]
-fn validation_output_summary_survives_restore_sanitized() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger = tmp.path().join("sessions.json");
-    let store = persistent_store(ledger.clone());
-    let session = store.start_session(None, Some("validation output".to_string()));
-    let start = store.record_tool_call_started(
-        Some(&session.session_id),
-        SessionTransport::Api,
-        "cargo_check",
-        &json!({"project": "agent:eval:demo"}),
-        crate::tool_runtime::sessions::session_tool_contract("cargo_check"),
-    );
-    store.record_tool_call_finished(
-        start,
-        false,
-        &json!({
-            "exit_code": 101,
-            "stdout": "full stdout body must not persist",
-            "stderr": "full stderr body must not persist",
-            "stdout_tail": "token=supersecret\nsafe stdout line\n",
-            "stderr_tail": "Authorization: Bearer supersecret\nerror[E0308]: mismatched types\n --> src/lib.rs:12:5\n",
-            "stdout_truncated": false,
-            "stderr_truncated": false,
-        }),
-        Some("tool failed"),
-        None,
-    );
-
-    let restored = flush_and_restore(&store, ledger);
-    let summary = restored.summary(&session.session_id, Some(10)).unwrap();
-    let finished = summary
-        .events
-        .iter()
-        .find(|event| event.kind == "tool_call_finished")
-        .unwrap();
-    let output_summary = finished.validation_output_summary.as_ref().unwrap();
-    let stdout_excerpt = output_summary["stdout_tail_excerpt"].as_str().unwrap();
-    let stderr_excerpt = output_summary["stderr_tail_excerpt"].as_str().unwrap();
-
-    assert_eq!(output_summary["tool_name"], "cargo_check");
-    assert!(stdout_excerpt.contains("safe stdout line"));
-    assert!(stderr_excerpt.contains("error[E0308]"));
-    assert!(stderr_excerpt.contains("--> src/lib.rs:12:5"));
-    for leaked in [
-        "full stdout body must not persist",
-        "full stderr body must not persist",
-        "token=supersecret",
-        "Authorization: Bearer supersecret",
-    ] {
-        assert!(
-            !serde_json::to_string(output_summary)
-                .unwrap()
-                .contains(leaked),
-            "restored validation_output_summary leaked {leaked}: {output_summary}"
-        );
-    }
-    assert!(stdout_excerpt.chars().count() <= MAX_VALIDATION_EXCERPT_CHARS);
-    assert!(stderr_excerpt.chars().count() <= MAX_VALIDATION_EXCERPT_CHARS);
-    assert_eq!(output_summary["stdout_truncated"], true);
-    assert_eq!(output_summary["stderr_truncated"], true);
-}
-
-#[test]
-fn malicious_persisted_validation_output_summary_is_resanitized_on_restore() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger = tmp.path().join("sessions.json");
-    let store = persistent_store(ledger.clone());
-    let session = store.start_session(None, Some("malicious validation".to_string()));
-    for tool_name in ["cargo_check", "run_shell"] {
-        let start = store.record_tool_call_started(
-            Some(&session.session_id),
-            SessionTransport::Api,
-            tool_name,
-            &json!({"project": "agent:eval:demo"}),
-            crate::tool_runtime::sessions::session_tool_contract(tool_name),
-        );
-        store.record_tool_call_finished(
-            start,
-            false,
-            &json!({"exit_code": 101}),
-            Some("tool failed"),
-            None,
-        );
-    }
-
-    store.flush_persistence();
-    let mut ledger_value: Value =
-        serde_json::from_str(&std::fs::read_to_string(&ledger).unwrap()).unwrap();
-    let events = ledger_value["sessions"][0]["events"]
-        .as_array_mut()
-        .unwrap();
-    for event in events {
-        if event["kind"] != "tool_call_finished" {
-            continue;
-        }
-        let tool_name = event["tool_name"].clone();
-        event["validation_output_summary"] = json!({
-            "tool_name": tool_name,
-            "stdout_tail_excerpt": format!(
-                "token=abc\nsecret=abc\npassword=abc\napi_key=abc\n{}STDOUT_SAFE_END",
-                "x".repeat(MAX_VALIDATION_EXCERPT_CHARS + 64)
-            ),
-            "stderr_tail_excerpt": format!(
-                "authorization: basic abc\nbearer abc\nprivate key abc\naccess key abc\n{}STDERR_SAFE_END",
-                "y".repeat(MAX_VALIDATION_EXCERPT_CHARS + 64)
-            ),
-            "stdout_truncated": false,
-            "stderr_truncated": false,
-            "max_excerpt_chars": 999999,
-        });
-    }
-    std::fs::write(&ledger, serde_json::to_vec_pretty(&ledger_value).unwrap()).unwrap();
-
-    let restored = flush_and_restore(&store, ledger);
-    let summary = restored.summary(&session.session_id, Some(10)).unwrap();
-    let cargo_finished = summary
-        .events
-        .iter()
-        .find(|event| event.kind == "tool_call_finished" && event.tool_name == "cargo_check")
-        .unwrap();
-    let run_shell_finished = summary
-        .events
-        .iter()
-        .find(|event| event.kind == "tool_call_finished" && event.tool_name == "run_shell")
-        .unwrap();
-    for output_summary in [
-        cargo_finished.validation_output_summary.as_ref().unwrap(),
-        run_shell_finished
-            .validation_output_summary
-            .as_ref()
-            .unwrap(),
-    ] {
-        let stdout_excerpt = output_summary["stdout_tail_excerpt"].as_str().unwrap();
-        let stderr_excerpt = output_summary["stderr_tail_excerpt"].as_str().unwrap();
-        let serialized = serde_json::to_string(output_summary).unwrap();
-
-        assert!(stdout_excerpt.contains("STDOUT_SAFE_END"));
-        assert!(stderr_excerpt.contains("STDERR_SAFE_END"));
-        assert!(stdout_excerpt.chars().count() <= MAX_VALIDATION_EXCERPT_CHARS);
-        assert!(stderr_excerpt.chars().count() <= MAX_VALIDATION_EXCERPT_CHARS);
-        assert_eq!(
-            output_summary["max_excerpt_chars"],
-            MAX_VALIDATION_EXCERPT_CHARS
-        );
-        assert_eq!(output_summary["stdout_truncated"], true);
-        assert_eq!(output_summary["stderr_truncated"], true);
-        for leaked in [
-            "token=abc",
-            "secret=abc",
-            "password=abc",
-            "api_key=abc",
-            "authorization: basic abc",
-            "bearer abc",
-            "private key abc",
-            "access key abc",
-        ] {
-            assert!(
-                !serialized.contains(leaked),
-                "restored validation_output_summary leaked {leaked}: {serialized}"
-            );
-        }
-    }
-}
-
-#[test]
-fn tool_call_start_and_finish_share_one_call_id() {
-    let store = SessionStore::new_in_memory(10, 20);
-    let session = store.start_session(
-        Some("agent:eval:demo".to_string()),
-        Some("correlation".to_string()),
-    );
-    let start = store
-        .record_tool_call_started(
-            Some(&session.session_id),
-            SessionTransport::Api,
-            "read_file",
-            &json!({"project": "agent:eval:demo", "path": "src/lib.rs"}),
-            crate::tool_runtime::sessions::session_tool_contract("read_file"),
-        )
-        .expect("start recorded");
-    let call_id = start.call_id.clone();
-    assert!(call_id.starts_with("wc_call_"));
-    store.record_tool_call_finished(
-        Some(start),
-        true,
-        &json!({"content": "omitted"}),
-        None,
-        None,
-    );
-    let summary = store.summary(&session.session_id, Some(20)).unwrap();
-    let tool_events = summary
-        .events
-        .iter()
-        .filter(|event| event.kind.starts_with("tool_call_"))
-        .collect::<Vec<_>>();
-    assert_eq!(tool_events.len(), 2);
-    assert_eq!(tool_events[0].call_id.as_deref(), Some(call_id.as_str()));
-    assert_eq!(tool_events[1].call_id.as_deref(), Some(call_id.as_str()));
-}
-
-#[test]
 fn legacy_session_events_without_call_id_restore() {
     let tmp = tempfile::tempdir().unwrap();
     let ledger = tmp.path().join("sessions.json");
@@ -1608,93 +1068,6 @@ fn legacy_session_events_without_call_id_restore() {
     assert_eq!(detail.activity.len(), 1);
     assert_eq!(detail.activity[0].state, "succeeded");
     assert!(!detail.running_call);
-}
-
-#[test]
-fn reused_logical_invocation_metadata_does_not_suppress_ambiguous_evidence() {
-    let store = SessionStore::new_in_memory(10, 20);
-    let session = store.start_session(Some("agent:eval:demo".to_string()), None);
-    let mut business = ToolCallRecorderMetadata::default();
-    business.assign_logical_invocation();
-    business.mark_business_execution();
-
-    for path in ["src/one.rs", "src/two.rs"] {
-        let start = store.record_tool_call_started_with_metadata(
-            Some(&session.session_id),
-            SessionTransport::Api,
-            "read_file",
-            &json!({"project": "agent:eval:demo", "path": path}),
-            Some("agent:eval:demo".to_string()),
-            business.clone(),
-            crate::tool_runtime::sessions::session_tool_contract("read_file"),
-        );
-        store.record_tool_call_finished(start, true, &json!({"content": "omitted"}), None, None);
-    }
-
-    let summary = store.summary(&session.session_id, Some(20)).unwrap();
-    assert_eq!(
-        summary.counts.tool_calls, 2,
-        "a duplicated valid-looking business correlation is ambiguous and must stay conservative"
-    );
-    let canonical = super::events::canonical_tool_call_finished_events(&summary.events);
-    assert_eq!(canonical.len(), 2);
-}
-
-fn record_console_tool(
-    store: &SessionStore,
-    session_id: &str,
-    project: &str,
-    tool: &str,
-    input: Value,
-    success: bool,
-    output: Value,
-) {
-    let start = store.record_tool_call_started(
-        Some(session_id),
-        SessionTransport::Api,
-        tool,
-        &input,
-        crate::tool_runtime::sessions::session_tool_contract(tool),
-    );
-    store.record_tool_call_finished(
-        start,
-        success,
-        &output,
-        (!success).then_some("tool failed"),
-        (!success).then_some("validation_failed"),
-    );
-    debug_assert_eq!(input.get("project").and_then(Value::as_str), Some(project));
-}
-
-fn record_console_validation(
-    store: &SessionStore,
-    session_id: &str,
-    project: &str,
-    success: bool,
-    tests_run_count: u64,
-) {
-    record_console_tool(
-        store,
-        session_id,
-        project,
-        "cargo_test",
-        json!({"project": project}),
-        success,
-        json!({
-            "exit_code": if success { 0 } else { 101 },
-            "execution_state": "completed",
-            "command_started": true,
-            "command_completed": true,
-            "tests_detected": true,
-            "tests_run_count": tests_run_count,
-            "zero_tests_run": tests_run_count == 0,
-            "stdout_tail": if success { "test result: ok" } else { "test result: FAILED" },
-            "stderr_tail": "",
-            "stdout_lines": 1,
-            "stderr_lines": 0,
-            "failure_kind": if success { Value::Null } else { json!("test_failure") }
-        }),
-    );
 }
 
 #[test]
@@ -2969,40 +2342,6 @@ fn console_title_and_progress_redact_json_wrapped_private_paths() {
 }
 
 #[test]
-fn legacy_session_events_without_validation_output_summary_restore() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger = tmp.path().join("sessions.json");
-    let store = persistent_store(ledger.clone());
-    let session = store.start_session(None, Some("legacy validation".to_string()));
-    let start = store.record_tool_call_started(
-        Some(&session.session_id),
-        SessionTransport::Api,
-        "cargo_check",
-        &json!({"project": "agent:eval:demo"}),
-        crate::tool_runtime::sessions::session_tool_contract("cargo_check"),
-    );
-    store.record_tool_call_finished(start, true, &json!({"exit_code": 0}), None, None);
-
-    store.flush_persistence();
-    let ledger_text = std::fs::read_to_string(&ledger).unwrap();
-    assert!(
-        !ledger_text.contains("validation_output_summary"),
-        "legacy fixture should omit validation_output_summary: {ledger_text}"
-    );
-    let restored = flush_and_restore(&store, ledger);
-    let summary = restored.summary(&session.session_id, Some(10)).unwrap();
-    let finished = summary
-        .events
-        .iter()
-        .find(|event| event.kind == "tool_call_finished")
-        .unwrap();
-
-    assert_eq!(summary.counts.tool_calls, 1);
-    assert_eq!(finished.tool_name, "cargo_check");
-    assert!(finished.validation_output_summary.is_none());
-}
-
-#[test]
 fn failure_history_legacy_event_without_effect_evidence_restores_conservatively() {
     let tmp = tempfile::tempdir().unwrap();
     let ledger = tmp.path().join("sessions.json");
@@ -3057,41 +2396,6 @@ fn failure_history_legacy_event_without_effect_evidence_restores_conservatively(
     .tool_failures;
     assert_eq!(projected["historical_non_actionable_count"], 0);
     assert_eq!(projected["actionable_unexpected_count"], 1);
-}
-
-#[test]
-fn legacy_session_events_without_observed_paths_restore_with_empty_evidence() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger = tmp.path().join("sessions.json");
-    let store = persistent_store(ledger.clone());
-    let session = store.start_session(None, Some("legacy exploration".to_string()));
-    let start = store.record_tool_call_started(
-        Some(&session.session_id),
-        SessionTransport::Api,
-        "read_file",
-        &json!({"project": "demo", "path": "src/legacy.rs"}),
-        crate::tool_runtime::sessions::session_tool_contract("read_file"),
-    );
-    store.record_tool_call_finished(start, true, &json!({"content": "omitted"}), None, None);
-    store.flush_persistence();
-
-    let mut ledger_value: Value =
-        serde_json::from_str(&std::fs::read_to_string(&ledger).unwrap()).unwrap();
-    for event in ledger_value["sessions"][0]["events"]
-        .as_array_mut()
-        .unwrap()
-    {
-        event.as_object_mut().unwrap().remove("observed_paths");
-    }
-    std::fs::write(&ledger, serde_json::to_vec_pretty(&ledger_value).unwrap()).unwrap();
-
-    let restored = SessionStore::with_persistence(ledger, 10, 10);
-    let summary = restored.summary(&session.session_id, Some(10)).unwrap();
-    assert!(summary
-        .events
-        .iter()
-        .all(|event| event.observed_paths.is_empty()));
-    assert_eq!(restored.status().restored_sessions, 1);
 }
 
 #[test]
@@ -3352,63 +2656,6 @@ fn search_project_texts_nested_patterns_are_removed_from_session_input_summary()
 }
 
 #[test]
-fn resolved_message_survives_restore() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger = tmp.path().join("sessions.json");
-    let store = persistent_store(ledger.clone());
-    let session = store.start_session(None, None);
-    let message = post_message(
-        &store,
-        &session.session_id,
-        SessionMessageKind::Todo,
-        "finish persistence tests",
-    );
-    store
-        .resolve_message(
-            &session.session_id,
-            &message.message_id,
-            Some("covered".to_string()),
-        )
-        .unwrap();
-
-    let restored = flush_and_restore(&store, ledger);
-    let messages = restored
-        .list_messages(
-            &session.session_id,
-            ListSessionMessagesFilter {
-                kind: Some(SessionMessageKind::Todo),
-                status: Some(SessionMessageStatus::Resolved),
-                message_id: None,
-                reply_to: None,
-                limit: Some(10),
-            },
-        )
-        .unwrap();
-    assert_eq!(messages.len(), 1);
-    assert_eq!(messages[0].status, SessionMessageStatus::Resolved);
-    assert_eq!(messages[0].resolution.as_deref(), Some("covered"));
-    assert!(messages[0].resolved_at.is_some());
-}
-
-#[test]
-fn corrupted_ledger_does_not_panic() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger = tmp.path().join("sessions.json");
-    std::fs::write(&ledger, "{not valid json").unwrap();
-
-    let store = persistent_store(ledger);
-    let status = store.status();
-    assert_eq!(status.persistence, "enabled");
-    assert_eq!(status.restored_sessions, 0);
-    assert!(status
-        .last_persist_error
-        .as_deref()
-        .unwrap()
-        .contains("restore_failed"));
-    assert!(store.summary("wc_sess_missing", None).is_none());
-}
-
-#[test]
 fn persistence_snapshot_shares_payload_and_stays_stable_across_message_cow() {
     let tmp = tempfile::tempdir().unwrap();
     let ledger = tmp.path().join("sessions.json");
@@ -3607,1509 +2854,6 @@ fn post_message(
             priority: SessionMessagePriority::Normal,
         })
         .unwrap()
-}
-
-#[test]
-fn list_session_messages_filters_and_clamps_limit() {
-    let store = SessionStore::default();
-    let session = store.start_session(None, None);
-    post_message(
-        &store,
-        &session.session_id,
-        SessionMessageKind::Guidance,
-        "g1",
-    );
-    post_message(
-        &store,
-        &session.session_id,
-        SessionMessageKind::Progress,
-        "p1",
-    );
-    post_message(&store, &session.session_id, SessionMessageKind::Risk, "r1");
-
-    let guidance = store
-        .list_messages(
-            &session.session_id,
-            ListSessionMessagesFilter {
-                kind: Some(SessionMessageKind::Guidance),
-                status: None,
-                message_id: None,
-                reply_to: None,
-                limit: None,
-            },
-        )
-        .unwrap();
-    assert_eq!(guidance.len(), 1);
-    assert_eq!(guidance[0].kind, SessionMessageKind::Guidance);
-
-    let open = store
-        .list_messages(
-            &session.session_id,
-            ListSessionMessagesFilter {
-                kind: None,
-                status: Some(SessionMessageStatus::Open),
-                message_id: None,
-                reply_to: None,
-                limit: Some(usize::MAX),
-            },
-        )
-        .unwrap();
-    assert_eq!(open.len(), 3);
-    assert_eq!(open[0].message, "r1");
-}
-
-#[test]
-fn session_message_unknown_errors_are_explicit() {
-    let store = SessionStore::default();
-    let session = store.start_session(None, None);
-    let unknown_session = store.post_message(PostSessionMessageInput {
-        session_id: "wc_sess_missing".to_string(),
-        kind: SessionMessageKind::Note,
-        message: "hello".to_string(),
-        tags: Vec::new(),
-        reply_to: None,
-        priority: SessionMessagePriority::Normal,
-    });
-    assert!(matches!(
-        unknown_session,
-        Err(SessionMessageError::UnknownSession)
-    ));
-
-    let unknown_message = store.resolve_message(&session.session_id, "wc_msg_missing", None);
-    assert!(matches!(
-        unknown_message,
-        Err(SessionMessageError::UnknownMessage)
-    ));
-}
-
-#[test]
-fn session_summary_includes_bounded_message_summary() {
-    let store = SessionStore::default();
-    let session = store.start_session(None, None);
-    post_message(
-        &store,
-        &session.session_id,
-        SessionMessageKind::Guidance,
-        "g1",
-    );
-    post_message(
-        &store,
-        &session.session_id,
-        SessionMessageKind::Progress,
-        "p1",
-    );
-    post_message(&store, &session.session_id, SessionMessageKind::Risk, "r1");
-    post_message(&store, &session.session_id, SessionMessageKind::Todo, "t1");
-
-    let summary = store.summary(&session.session_id, Some(50)).unwrap();
-    assert_eq!(summary.messages.total, 4);
-    assert_eq!(summary.messages.open, 4);
-    assert_eq!(summary.messages.pending_guidance, 1);
-    assert_eq!(summary.messages.open_risks, 1);
-    assert_eq!(summary.messages.open_todos, 1);
-    assert_eq!(summary.messages.recent_progress.len(), 1);
-    assert!(serde_json::to_value(summary)
-        .unwrap()
-        .get("messages")
-        .is_some());
-}
-
-/// Create-entry funnel: convenience wrappers must produce the same shape as
-/// the authoritative `start_session_with_options` path.
-#[test]
-fn start_session_wrappers_funnel_to_single_create_entry() {
-    let store = SessionStore::default();
-
-    let via_start = store.start_session(Some("proj-a".to_string()), Some("t1".to_string()));
-    let via_guards = store.start_session_with_guards(
-        Some("proj-a".to_string()),
-        Some("t2".to_string()),
-        SessionMode::Normal,
-        SessionGuards::default(),
-    );
-    let via_options = store
-        .start_session_with_options(SessionCreateOptions::new(
-            Some("proj-a".to_string()),
-            Some("t3".to_string()),
-            SessionMode::Normal,
-            SessionGuards::default(),
-        ))
-        .unwrap();
-    let via_read_only = store.start_session_with_guards(
-        Some("proj-a".to_string()),
-        Some("ro".to_string()),
-        SessionMode::ReadOnly,
-        SessionGuards::default(),
-    );
-
-    for summary in [&via_start, &via_guards, &via_options, &via_read_only] {
-        assert!(summary.session_id.starts_with("wc_sess_"));
-        assert!(store.contains_session(&summary.session_id));
-        assert_eq!(summary.project.as_deref(), Some("proj-a"));
-    }
-    assert_eq!(via_start.mode, SessionMode::Normal);
-    assert!(!via_start.guards.deny_write_tools);
-    assert!(!via_start.guards.deny_shell_tools);
-    assert_eq!(via_read_only.mode, SessionMode::ReadOnly);
-    assert!(via_read_only.guards.deny_write_tools);
-    assert!(via_read_only.guards.deny_shell_tools);
-    assert!(store
-        .guard_denial(
-            &via_read_only.session_id,
-            crate::tool_runtime::sessions::session_tool_contract("write_project_file")
-        )
-        .is_some());
-    assert!(store
-        .guard_denial(
-            &via_start.session_id,
-            crate::tool_runtime::sessions::session_tool_contract("write_project_file")
-        )
-        .is_none());
-}
-
-/// Unknown sessions never accept events or messages, and are not recreated.
-#[test]
-fn unknown_session_mutations_do_not_recreate_session() {
-    let store = SessionStore::default();
-    let missing = "wc_sess_does_not_exist";
-
-    assert!(!store.contains_session(missing));
-    assert!(store.summary(missing, None).is_none());
-    assert!(store
-        .record_tool_call_started(
-            Some(missing),
-            SessionTransport::Api,
-            "read_file",
-            &json!({"project": "demo", "path": "a.rs"}),
-            crate::tool_runtime::sessions::session_tool_contract("read_file"),
-        )
-        .is_none());
-    assert!(!store.contains_session(missing));
-
-    let post = store.post_message(PostSessionMessageInput {
-        session_id: missing.to_string(),
-        kind: SessionMessageKind::Note,
-        message: "nope".to_string(),
-        tags: Vec::new(),
-        reply_to: None,
-        priority: SessionMessagePriority::Normal,
-    });
-    assert!(matches!(post, Err(SessionMessageError::UnknownSession)));
-    assert!(!store.contains_session(missing));
-
-    let resolve = store.resolve_message(missing, "wc_msg_x", None);
-    assert!(matches!(resolve, Err(SessionMessageError::UnknownSession)));
-}
-
-/// Evicted (capacity-bound) sessions stay gone: events must not revive them.
-#[test]
-fn evicted_session_is_not_reactivated_by_events_or_messages() {
-    let store = SessionStore::new(1, 10);
-    let first = store.start_session(None, Some("first".to_string()));
-    let second = store.start_session(None, Some("second".to_string()));
-
-    assert!(!store.contains_session(&first.session_id));
-    assert!(store.contains_session(&second.session_id));
-
-    assert!(store
-        .record_tool_call_started(
-            Some(&first.session_id),
-            SessionTransport::Api,
-            "read_file",
-            &json!({"project": "demo", "path": "a.rs"}),
-            crate::tool_runtime::sessions::session_tool_contract("read_file"),
-        )
-        .is_none());
-    assert!(!store.contains_session(&first.session_id));
-    assert!(store.summary(&first.session_id, None).is_none());
-
-    let post = store.post_message(PostSessionMessageInput {
-        session_id: first.session_id.clone(),
-        kind: SessionMessageKind::Note,
-        message: "revive?".to_string(),
-        tags: Vec::new(),
-        reply_to: None,
-        priority: SessionMessagePriority::Normal,
-    });
-    assert!(matches!(post, Err(SessionMessageError::UnknownSession)));
-    assert!(!store.contains_session(&first.session_id));
-}
-
-#[test]
-fn session_message_create_list_and_resolve_contract() {
-    let store = SessionStore::default();
-    let session = store.start_session(None, None);
-    let posted = store
-        .post_message(PostSessionMessageInput {
-            session_id: session.session_id.clone(),
-            kind: SessionMessageKind::Todo,
-            message: "do the thing".to_string(),
-            tags: vec!["work".to_string(), "constraint".to_string()],
-            reply_to: None,
-            priority: SessionMessagePriority::High,
-        })
-        .unwrap();
-    assert!(posted.message_id.starts_with(MESSAGE_ID_PREFIX));
-    assert_eq!(posted.session_id, session.session_id);
-    assert_eq!(posted.kind, SessionMessageKind::Todo);
-    assert_eq!(posted.status, SessionMessageStatus::Open);
-    assert_eq!(posted.priority, SessionMessagePriority::High);
-    assert_eq!(posted.message, "do the thing");
-    assert_eq!(posted.tags, vec!["work", "constraint"]);
-
-    let listed = store
-        .list_messages(
-            &session.session_id,
-            ListSessionMessagesFilter {
-                kind: Some(SessionMessageKind::Todo),
-                status: Some(SessionMessageStatus::Open),
-                message_id: None,
-                reply_to: None,
-                limit: Some(10),
-            },
-        )
-        .unwrap();
-    assert_eq!(listed.len(), 1);
-    assert_eq!(listed[0].message_id, posted.message_id);
-
-    let resolved = store
-        .resolve_message(
-            &session.session_id,
-            &posted.message_id,
-            Some("shipped".to_string()),
-        )
-        .unwrap();
-    assert_eq!(resolved.status, SessionMessageStatus::Resolved);
-    assert_eq!(resolved.resolution.as_deref(), Some("shipped"));
-    let first_resolved_at = resolved.resolved_at.expect("resolved_at set");
-
-    let idempotent = store
-        .resolve_message(&session.session_id, &posted.message_id, None)
-        .unwrap();
-    assert_eq!(idempotent.status, SessionMessageStatus::Resolved);
-    assert_eq!(idempotent.resolved_at, Some(first_resolved_at));
-    assert_eq!(idempotent.resolution.as_deref(), Some("shipped"));
-
-    let updated = store
-        .resolve_message(
-            &session.session_id,
-            &posted.message_id,
-            Some("still done".to_string()),
-        )
-        .unwrap();
-    assert_eq!(updated.status, SessionMessageStatus::Resolved);
-    assert_eq!(updated.resolved_at, Some(first_resolved_at));
-    assert_eq!(updated.resolution.as_deref(), Some("still done"));
-
-    let open = store
-        .list_messages(
-            &session.session_id,
-            ListSessionMessagesFilter {
-                kind: None,
-                status: Some(SessionMessageStatus::Open),
-                message_id: None,
-                reply_to: None,
-                limit: None,
-            },
-        )
-        .unwrap();
-    assert!(open.is_empty());
-}
-
-#[test]
-fn wrapper_resolution_requires_ack_rejects_todo_and_replays_idempotently() {
-    let store = SessionStore::default();
-    let session = store.start_session(None, None);
-    let guidance = store
-        .post_message_with_ack(
-            PostSessionMessageInput {
-                session_id: session.session_id.clone(),
-                kind: SessionMessageKind::Guidance,
-                message: "apply the reviewed direction".to_string(),
-                tags: Vec::new(),
-                reply_to: None,
-                priority: SessionMessagePriority::High,
-            },
-            true,
-        )
-        .unwrap();
-
-    let missing_ack = store.resolve_message_from_wrapper(
-        &session.session_id,
-        &guidance.message_id,
-        "handled".to_string(),
-        false,
-    );
-    assert!(matches!(
-        missing_ack,
-        Err(SessionMessageError::InvalidInput(_))
-    ));
-    let still_open = store
-        .list_messages(
-            &session.session_id,
-            ListSessionMessagesFilter {
-                message_id: Some(guidance.message_id.clone()),
-                ..Default::default()
-            },
-        )
-        .unwrap();
-    assert_eq!(still_open[0].status, SessionMessageStatus::Open);
-
-    let resolved = store
-        .resolve_message_from_wrapper(
-            &session.session_id,
-            &guidance.message_id,
-            "handled".to_string(),
-            true,
-        )
-        .unwrap();
-    assert_eq!(resolved.status, SessionMessageStatus::Resolved);
-    assert_eq!(resolved.resolution.as_deref(), Some("handled"));
-    let resolved_at = resolved.resolved_at;
-
-    let replay = store
-        .resolve_message_from_wrapper(
-            &session.session_id,
-            &guidance.message_id,
-            "handled".to_string(),
-            false,
-        )
-        .unwrap();
-    assert_eq!(replay.resolved_at, resolved_at);
-    let conflict = store.resolve_message_from_wrapper(
-        &session.session_id,
-        &guidance.message_id,
-        "different completion".to_string(),
-        true,
-    );
-    assert!(matches!(
-        conflict,
-        Err(SessionMessageError::IdempotencyConflict)
-    ));
-
-    let todo = store
-        .post_message(PostSessionMessageInput {
-            session_id: session.session_id.clone(),
-            kind: SessionMessageKind::Todo,
-            message: "finish the delegated task".to_string(),
-            tags: Vec::new(),
-            reply_to: None,
-            priority: SessionMessagePriority::Normal,
-        })
-        .unwrap();
-    let todo_resolution = store.resolve_message_from_wrapper(
-        &session.session_id,
-        &todo.message_id,
-        "done".to_string(),
-        true,
-    );
-    assert!(matches!(
-        todo_resolution,
-        Err(SessionMessageError::InvalidInput(_))
-    ));
-    let retained_todo = store
-        .list_messages(
-            &session.session_id,
-            ListSessionMessagesFilter {
-                message_id: Some(todo.message_id),
-                ..Default::default()
-            },
-        )
-        .unwrap();
-    assert_eq!(retained_todo[0].status, SessionMessageStatus::Open);
-}
-
-#[test]
-fn read_only_guards_block_write_and_shell_classifications() {
-    let store = SessionStore::default();
-    let normal =
-        store.start_session_with_guards(None, None, SessionMode::Normal, SessionGuards::default());
-    let read_only = store.start_session_with_guards(
-        None,
-        None,
-        SessionMode::ReadOnly,
-        SessionGuards::default(),
-    );
-    assert!(store
-        .guard_denial(
-            &normal.session_id,
-            crate::tool_runtime::sessions::session_tool_contract("write_project_file")
-        )
-        .is_none());
-    assert!(store
-        .guard_denial(
-            &normal.session_id,
-            crate::tool_runtime::sessions::session_tool_contract("run_shell")
-        )
-        .is_none());
-
-    let write_denial = store
-        .guard_denial(
-            &read_only.session_id,
-            crate::tool_runtime::sessions::session_tool_contract("write_project_file"),
-        )
-        .expect("write denied");
-    assert_eq!(write_denial.guard, "deny_write_tools");
-    assert_eq!(write_denial.mode, SessionMode::ReadOnly);
-
-    let shell_denial = store
-        .guard_denial(
-            &read_only.session_id,
-            crate::tool_runtime::sessions::session_tool_contract("run_shell"),
-        )
-        .expect("shell denied");
-    assert_eq!(shell_denial.guard, "deny_shell_tools");
-
-    // Reads remain allowed under read_only.
-    assert!(store
-        .guard_denial(
-            &read_only.session_id,
-            crate::tool_runtime::sessions::session_tool_contract("read_file")
-        )
-        .is_none());
-}
-
-#[test]
-fn ledger_round_trip_preserves_session_state_events_and_messages() {
-    let dir = tempfile::tempdir().unwrap();
-    let ledger = dir.path().join("sessions.json");
-    let store = SessionStore::with_persistence(&ledger, 10, 50);
-    let session = store.start_session_with_guards(
-        Some("proj".to_string()),
-        Some("persist".to_string()),
-        SessionMode::ReadOnly,
-        SessionGuards::default(),
-    );
-    let start = store
-        .record_tool_call_started(
-            Some(&session.session_id),
-            SessionTransport::Api,
-            "read_file",
-            &json!({"project": "proj", "path": "src/lib.rs"}),
-            crate::tool_runtime::sessions::session_tool_contract("read_file"),
-        )
-        .unwrap();
-    store.record_tool_call_finished(Some(start), true, &json!({}), None, None);
-    post_message(
-        &store,
-        &session.session_id,
-        SessionMessageKind::Guidance,
-        "keep OpenAPI operation count stable",
-    );
-    post_message(
-        &store,
-        &session.session_id,
-        SessionMessageKind::Progress,
-        "checkpoint",
-    );
-
-    store.flush_persistence();
-    let restored = SessionStore::with_persistence(&ledger, 10, 50);
-    let summary = restored.summary(&session.session_id, Some(20)).unwrap();
-    assert_eq!(summary.project.as_deref(), Some("proj"));
-    assert_eq!(summary.title.as_deref(), Some("persist"));
-    assert_eq!(summary.mode, SessionMode::ReadOnly);
-    assert!(summary.guards.deny_write_tools);
-    assert!(summary.guards.deny_shell_tools);
-    assert_eq!(summary.lifecycle, SessionLifecycle::Active);
-    assert_eq!(summary.counts.tool_calls, 1);
-    assert_eq!(summary.counts.succeeded, 1);
-    assert!(summary
-        .events
-        .iter()
-        .any(|event| event.kind == "tool_call_finished" && event.tool_name == "read_file"));
-
-    let messages = restored
-        .list_messages(&session.session_id, ListSessionMessagesFilter::default())
-        .unwrap();
-    assert_eq!(messages.len(), 2);
-    assert_eq!(messages[0].message, "checkpoint");
-    assert_eq!(messages[1].kind, SessionMessageKind::Guidance);
-    let discussion = restored
-        .discussion_summary(&session.session_id, Some(10))
-        .unwrap();
-    assert_eq!(discussion.counts.total, 2);
-    assert_eq!(discussion.counts.guidance, 1);
-    assert_eq!(discussion.counts.progress, 1);
-}
-
-// --- Workflow Session lifecycle: explicit Active/Closed authority ---
-
-#[test]
-fn new_session_defaults_lifecycle_to_active() {
-    let store = SessionStore::default();
-    let summary = store.start_session(None, Some("lifecycle".to_string()));
-    assert_eq!(summary.lifecycle, SessionLifecycle::Active);
-
-    // Summary JSON exposes lifecycle for observability.
-    let value = serde_json::to_value(&summary).unwrap();
-    assert_eq!(value["lifecycle"], "active");
-}
-
-#[test]
-fn persisted_ledger_writes_and_reads_lifecycle_active() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger = tmp.path().join("sessions.json");
-    let store = persistent_store(ledger.clone());
-    let session = store.start_session(Some("proj".to_string()), Some("with lifecycle".to_string()));
-    assert_eq!(session.lifecycle, SessionLifecycle::Active);
-
-    store.flush_persistence();
-    let raw = std::fs::read_to_string(&ledger).unwrap();
-    let ledger_value: Value = serde_json::from_str(&raw).unwrap();
-    assert_eq!(ledger_value["version"], SESSION_LEDGER_VERSION);
-    assert_eq!(ledger_value["sessions"][0]["lifecycle"], "active");
-    assert_eq!(
-        ledger_value["sessions"][0]["session_id"].as_str().unwrap(),
-        session.session_id
-    );
-
-    let restored = flush_and_restore(&store, ledger);
-    let summary = restored.summary(&session.session_id, Some(10)).unwrap();
-    assert_eq!(summary.lifecycle, SessionLifecycle::Active);
-}
-
-#[test]
-fn pre_current_ledger_version_is_rejected_without_rewrite() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger_path = tmp.path().join("sessions.json");
-    let store = persistent_store(ledger_path.clone());
-    let session = store.start_session(Some("proj".to_string()), Some("current".to_string()));
-    store.flush_persistence();
-    drop(store);
-
-    let mut ledger: Value =
-        serde_json::from_str(&std::fs::read_to_string(&ledger_path).unwrap()).unwrap();
-    ledger["version"] = json!(1);
-    std::fs::write(&ledger_path, serde_json::to_vec_pretty(&ledger).unwrap()).unwrap();
-    let before = std::fs::read(&ledger_path).unwrap();
-
-    let restored = persistent_store(ledger_path.clone());
-    assert_eq!(restored.status().restored_sessions, 0);
-    assert!(restored.summary(&session.session_id, None).is_none());
-    assert!(restored
-        .status()
-        .last_persist_error
-        .as_deref()
-        .is_some_and(|error| error.contains("unsupported session ledger version 1")));
-    assert_eq!(std::fs::read(&ledger_path).unwrap(), before);
-}
-
-#[test]
-fn legacy_inspect_v2_row_is_discarded_without_affecting_valid_rows() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger_path = tmp.path().join("sessions.json");
-    let store = persistent_store(ledger_path.clone());
-    let retired = store.start_session(Some("proj".to_string()), Some("legacy inspect".to_string()));
-    let valid = store.start_session(Some("proj".to_string()), Some("normal".to_string()));
-    store.flush_persistence();
-    drop(store);
-
-    let mut ledger: Value =
-        serde_json::from_str(&std::fs::read_to_string(&ledger_path).unwrap()).unwrap();
-    let retired_record = ledger["sessions"]
-        .as_array_mut()
-        .unwrap()
-        .iter_mut()
-        .find(|record| record["session_id"] == retired.session_id)
-        .unwrap();
-    retired_record["mode"] = json!("inspect");
-    std::fs::write(&ledger_path, serde_json::to_vec_pretty(&ledger).unwrap()).unwrap();
-
-    let restored = persistent_store(ledger_path);
-    assert_eq!(restored.status().restored_sessions, 1);
-    assert!(restored.summary(&retired.session_id, None).is_none());
-    let valid_summary = restored.summary(&valid.session_id, None).unwrap();
-    assert_eq!(valid_summary.mode, SessionMode::Normal);
-    assert_eq!(restored.status().last_persist_error, None);
-}
-
-#[test]
-fn v2_missing_assignment_metadata_discards_only_that_row() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger_path = tmp.path().join("sessions.json");
-    let store = persistent_store(ledger_path.clone());
-    let bad = store.start_session(Some("proj".to_string()), Some("bad".to_string()));
-    let good = store.start_session(Some("proj".to_string()), Some("good".to_string()));
-    store.flush_persistence();
-    drop(store);
-
-    let mut ledger: Value =
-        serde_json::from_str(&std::fs::read_to_string(&ledger_path).unwrap()).unwrap();
-    let bad_record = ledger["sessions"]
-        .as_array_mut()
-        .unwrap()
-        .iter_mut()
-        .find(|record| record["session_id"] == bad.session_id)
-        .unwrap()
-        .as_object_mut()
-        .unwrap();
-    bad_record.remove("assignment_history_tracking_complete");
-    std::fs::write(&ledger_path, serde_json::to_vec_pretty(&ledger).unwrap()).unwrap();
-
-    let restored = persistent_store(ledger_path);
-    assert_eq!(restored.status().restored_sessions, 1);
-    assert!(restored.summary(&bad.session_id, None).is_none());
-    assert!(restored.summary(&good.session_id, None).is_some());
-    assert_eq!(restored.status().last_persist_error, None);
-}
-
-#[test]
-fn v2_event_and_message_shape_corruption_discards_only_affected_rows() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger_path = tmp.path().join("sessions.json");
-    let store = persistent_store(ledger_path.clone());
-    let bad_missing = store.start_session(
-        Some("proj".to_string()),
-        Some("missing correlation key".to_string()),
-    );
-    let bad_retired = store.start_session(
-        Some("proj".to_string()),
-        Some("retired event field".to_string()),
-    );
-    let bad_noncanonical_correlation = store.start_session(
-        Some("proj".to_string()),
-        Some("noncanonical correlation values".to_string()),
-    );
-    let bad_message = store.start_session(
-        Some("proj".to_string()),
-        Some("unknown message field".to_string()),
-    );
-    let good = store.start_session(Some("proj".to_string()), Some("good".to_string()));
-
-    for session_id in [
-        &bad_missing.session_id,
-        &bad_retired.session_id,
-        &bad_noncanonical_correlation.session_id,
-    ] {
-        let mut metadata = ToolCallRecorderMetadata::default();
-        metadata.assign_logical_invocation();
-        let start = store.record_tool_call_started_with_metadata(
-            Some(session_id),
-            SessionTransport::Api,
-            "read_file",
-            &json!({"project": "proj", "path": "src/lib.rs"}),
-            Some("proj".to_string()),
-            metadata,
-            crate::tool_runtime::sessions::session_tool_contract("read_file"),
-        );
-        store.record_tool_call_finished(start, true, &json!({"content": "omitted"}), None, None);
-    }
-    store
-        .post_message(PostSessionMessageInput {
-            session_id: bad_message.session_id.clone(),
-            kind: SessionMessageKind::Note,
-            message: "message row".to_string(),
-            tags: Vec::new(),
-            reply_to: None,
-            priority: SessionMessagePriority::Normal,
-        })
-        .unwrap();
-    store.flush_persistence();
-    drop(store);
-
-    let mut ledger: Value =
-        serde_json::from_str(&std::fs::read_to_string(&ledger_path).unwrap()).unwrap();
-    let records = ledger["sessions"].as_array_mut().unwrap();
-    let missing_record = records
-        .iter_mut()
-        .find(|record| record["session_id"] == bad_missing.session_id)
-        .unwrap();
-    let missing_event = missing_record["events"][0].as_object_mut().unwrap();
-    assert!(missing_event.contains_key("logical_invocation_id"));
-    assert!(missing_event.contains_key("logical_invocation_role"));
-    missing_event.remove("logical_invocation_role");
-
-    let retired_record = records
-        .iter_mut()
-        .find(|record| record["session_id"] == bad_retired.session_id)
-        .unwrap();
-    retired_record["events"][0]
-        .as_object_mut()
-        .unwrap()
-        .insert("allow_cross_project_session".to_string(), Value::Bool(true));
-
-    let noncanonical_record = records
-        .iter_mut()
-        .find(|record| record["session_id"] == bad_noncanonical_correlation.session_id)
-        .unwrap();
-    let noncanonical_event = noncanonical_record["events"][0].as_object_mut().unwrap();
-    noncanonical_event.insert("logical_invocation_id".to_string(), Value::Null);
-    noncanonical_event.insert("logical_invocation_role".to_string(), Value::Null);
-
-    let message_record = records
-        .iter_mut()
-        .find(|record| record["session_id"] == bad_message.session_id)
-        .unwrap();
-    message_record["messages"][0]
-        .as_object_mut()
-        .unwrap()
-        .insert(
-            "unsupported_development_field".to_string(),
-            Value::Bool(true),
-        );
-
-    std::fs::write(&ledger_path, serde_json::to_vec_pretty(&ledger).unwrap()).unwrap();
-    let restored = persistent_store(ledger_path);
-    assert_eq!(restored.status().restored_sessions, 1);
-    assert!(restored.summary(&bad_missing.session_id, None).is_none());
-    assert!(restored.summary(&bad_retired.session_id, None).is_none());
-    assert!(restored
-        .summary(&bad_noncanonical_correlation.session_id, None)
-        .is_none());
-    assert!(restored.summary(&bad_message.session_id, None).is_none());
-    assert!(restored.summary(&good.session_id, None).is_some());
-    assert_eq!(restored.status().last_persist_error, None);
-}
-
-#[test]
-fn current_ledger_rejects_retired_top_level_binding_member() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger_path = tmp.path().join("sessions.json");
-    let store = persistent_store(ledger_path.clone());
-    let session = store.start_session(Some("proj".to_string()), Some("canonical".to_string()));
-    store.flush_persistence();
-    drop(store);
-
-    let mut ledger: Value =
-        serde_json::from_str(&std::fs::read_to_string(&ledger_path).unwrap()).unwrap();
-    ledger["durable_current_bindings"] = json!([]);
-    std::fs::write(&ledger_path, serde_json::to_vec_pretty(&ledger).unwrap()).unwrap();
-
-    let restored = persistent_store(ledger_path);
-    assert_eq!(restored.status().restored_sessions, 0);
-    assert!(restored.summary(&session.session_id, None).is_none());
-    assert!(restored
-        .status()
-        .last_persist_error
-        .as_deref()
-        .is_some_and(|error| error.contains("invalid v2 session ledger")));
-}
-
-#[test]
-fn v2_missing_or_removed_lifecycle_discards_only_that_row() {
-    for lifecycle in [None, Some(json!("archived"))] {
-        let tmp = tempfile::tempdir().unwrap();
-        let ledger_path = tmp.path().join("sessions.json");
-        let store = persistent_store(ledger_path.clone());
-        let bad = store.start_session(Some("proj".to_string()), Some("bad lifecycle".to_string()));
-        let good =
-            store.start_session(Some("proj".to_string()), Some("good lifecycle".to_string()));
-        store.flush_persistence();
-        drop(store);
-
-        let mut ledger: Value =
-            serde_json::from_str(&std::fs::read_to_string(&ledger_path).unwrap()).unwrap();
-        let bad_record = ledger["sessions"]
-            .as_array_mut()
-            .unwrap()
-            .iter_mut()
-            .find(|record| record["session_id"] == bad.session_id)
-            .unwrap()
-            .as_object_mut()
-            .unwrap();
-        match lifecycle.clone() {
-            Some(value) => {
-                bad_record.insert("lifecycle".to_string(), value);
-            }
-            None => {
-                bad_record.remove("lifecycle");
-            }
-        }
-        std::fs::write(&ledger_path, serde_json::to_vec_pretty(&ledger).unwrap()).unwrap();
-
-        let restored = persistent_store(ledger_path);
-        assert_eq!(restored.status().restored_sessions, 1);
-        assert!(restored.summary(&bad.session_id, None).is_none());
-        assert!(restored.summary(&good.session_id, None).is_some());
-    }
-}
-
-#[test]
-fn session_lifecycle_wire_values_are_snake_case() {
-    assert_eq!(
-        serde_json::to_value(SessionLifecycle::Active).unwrap(),
-        json!("active")
-    );
-    assert_eq!(
-        serde_json::to_value(SessionLifecycle::Closed).unwrap(),
-        json!("closed")
-    );
-    assert_eq!(
-        serde_json::from_value::<SessionLifecycle>(json!("active")).unwrap(),
-        SessionLifecycle::Active
-    );
-}
-
-// --- Workflow session lifecycle (Phase 2: explicit close) ---
-
-#[test]
-fn active_to_closed_succeeds_and_emits_session_closed_event() {
-    let store = SessionStore::default();
-    let session = store.start_session(None, Some("close me".to_string()));
-    assert_eq!(session.lifecycle, SessionLifecycle::Active);
-
-    let outcome = store.close_session(&session.session_id).unwrap();
-    assert!(!outcome.already_closed);
-    assert_eq!(outcome.summary.lifecycle, SessionLifecycle::Closed);
-    assert_eq!(outcome.summary.session_id, session.session_id);
-
-    let summary = store.summary(&session.session_id, Some(20)).unwrap();
-    assert_eq!(summary.lifecycle, SessionLifecycle::Closed);
-    let closed_events: Vec<_> = summary
-        .events
-        .iter()
-        .filter(|event| event.kind == "session_closed")
-        .collect();
-    assert_eq!(closed_events.len(), 1);
-    assert_eq!(closed_events[0].tool_name, "close_session");
-    assert_eq!(closed_events[0].status.as_deref(), Some("succeeded"));
-}
-
-#[test]
-fn closed_session_coldifies_payload_and_queries_without_reheating() {
-    let store = SessionStore::default();
-    let session = store.start_session(None, Some("cold history".to_string()));
-    post_message(
-        &store,
-        &session.session_id,
-        SessionMessageKind::Todo,
-        "retained closed message",
-    );
-    let start = store.record_tool_call_started(
-        Some(&session.session_id),
-        SessionTransport::Api,
-        "read_file",
-        &json!({"project": "demo", "path": "src/lib.rs"}),
-        crate::tool_runtime::sessions::session_tool_contract("read_file"),
-    );
-    store.record_tool_call_finished(start, true, &json!({"content": "retained"}), None, None);
-    assert!(store
-        .hot_payload_entry_count_for_test(&session.session_id)
-        .is_some_and(|entries| entries > 0));
-
-    let outcome = store.close_session(&session.session_id).unwrap();
-    assert!(!outcome.already_closed);
-    assert_eq!(
-        store.hot_payload_entry_count_for_test(&session.session_id),
-        None
-    );
-    assert!(store
-        .cold_payload_bytes_for_test(&session.session_id)
-        .is_some_and(|bytes| bytes > 0));
-
-    let other = store.start_session(None, Some("other cold history".to_string()));
-    store.close_session(&other.session_id).unwrap();
-    assert!(store
-        .cold_payload_bytes_for_test(&other.session_id)
-        .is_some());
-
-    let summary = store.summary(&session.session_id, Some(50)).unwrap();
-    assert_eq!(summary.lifecycle, SessionLifecycle::Closed);
-    assert!(summary
-        .events
-        .iter()
-        .any(|event| event.kind == "session_closed"));
-    let messages = store
-        .list_messages(&session.session_id, ListSessionMessagesFilter::default())
-        .unwrap();
-    assert_eq!(messages.len(), 1);
-    assert_eq!(messages[0].message, "retained closed message");
-    assert_eq!(
-        store.hot_payload_entry_count_for_test(&session.session_id),
-        None,
-        "cold queries must not install a hot SessionRecord"
-    );
-    assert_eq!(
-        store.hot_payload_entry_count_for_test(&other.session_id),
-        None,
-        "querying one cold Session must not heat another historical Session"
-    );
-
-    let query_start = store.record_tool_call_started(
-        Some(&session.session_id),
-        SessionTransport::Api,
-        "session_summary",
-        &json!({"session_id": session.session_id.clone()}),
-        crate::tool_runtime::sessions::session_tool_contract("session_summary"),
-    );
-    store.record_tool_call_finished(query_start, true, &json!({"success": true}), None, None);
-    assert_eq!(
-        store.hot_payload_entry_count_for_test(&session.session_id),
-        None,
-        "closed recorder events must rewrite the cold payload in place"
-    );
-
-    let repeated = store.close_session(&session.session_id).unwrap();
-    assert!(repeated.already_closed);
-    let summary = store.summary(&session.session_id, Some(100)).unwrap();
-    assert_eq!(
-        summary
-            .events
-            .iter()
-            .filter(|event| event.kind == "session_closed")
-            .count(),
-        1
-    );
-}
-
-#[test]
-fn close_cleanup_evidence_rewrites_cold_payload_without_reheating() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger = tmp.path().join("sessions.json");
-    let store = persistent_store(ledger.clone());
-    let session = store.start_session(None, Some("cold close evidence".to_string()));
-    store.close_session(&session.session_id).unwrap();
-    assert!(store
-        .cold_payload_bytes_for_test(&session.session_id)
-        .is_some());
-
-    store.record_session_close_persistent_shell_evidence(
-        &session.session_id,
-        "wc_shell_cold_close",
-        "closed",
-        "completed",
-        None,
-        false,
-    );
-    assert_eq!(
-        store.hot_payload_entry_count_for_test(&session.session_id),
-        None
-    );
-
-    let restored = flush_and_restore(&store, ledger);
-    assert!(restored
-        .cold_payload_bytes_for_test(&session.session_id)
-        .is_some());
-    let summary = restored.summary(&session.session_id, Some(50)).unwrap();
-    let evidence = summary
-        .events
-        .iter()
-        .find(|event| event.kind == "session_closed")
-        .and_then(|event| event.persistent_shell.as_ref())
-        .expect("close cleanup evidence must survive cold persistence");
-    assert_eq!(evidence.shell_id.as_deref(), Some("wc_shell_cold_close"));
-    assert_eq!(evidence.shell_state.as_deref(), Some("closed"));
-    assert_eq!(evidence.execution_state.as_deref(), Some("completed"));
-}
-
-#[test]
-fn closed_cold_round_trip_preserves_evidence_and_active_stays_hot() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger = tmp.path().join("sessions.json");
-    let store = SessionStore::with_persistence(&ledger, 10, 20);
-    let closed = store.start_session(Some("proj".to_string()), Some("historical".to_string()));
-    post_message(
-        &store,
-        &closed.session_id,
-        SessionMessageKind::Note,
-        "durable historical message",
-    );
-    let start = store.record_tool_call_started(
-        Some(&closed.session_id),
-        SessionTransport::Api,
-        "read_file",
-        &json!({"project": "proj", "path": "history.rs"}),
-        crate::tool_runtime::sessions::session_tool_contract("read_file"),
-    );
-    store.record_tool_call_finished(
-        start,
-        true,
-        &json!({"content": "history evidence"}),
-        None,
-        None,
-    );
-    store.close_session(&closed.session_id).unwrap();
-
-    let active = store.start_session(Some("proj".to_string()), Some("active".to_string()));
-    post_message(
-        &store,
-        &active.session_id,
-        SessionMessageKind::Progress,
-        "active message",
-    );
-    assert_eq!(
-        store.hot_payload_entry_count_for_test(&closed.session_id),
-        None
-    );
-    assert!(store
-        .hot_payload_entry_count_for_test(&active.session_id)
-        .is_some());
-    assert_eq!(store.cold_payload_bytes_for_test(&active.session_id), None);
-
-    let before_summary = store.summary(&closed.session_id, Some(100)).unwrap();
-    let before_messages = store
-        .list_messages(&closed.session_id, ListSessionMessagesFilter::default())
-        .unwrap();
-    store.flush_persistence();
-    let raw = std::fs::read_to_string(&ledger).unwrap();
-    let ledger_value: Value = serde_json::from_str(&raw).unwrap();
-    assert_eq!(ledger_value["version"], SESSION_LEDGER_VERSION);
-    let closed_json = ledger_value["sessions"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .find(|record| record["session_id"] == closed.session_id)
-        .unwrap();
-    assert_eq!(closed_json["lifecycle"], "closed");
-    assert!(closed_json["messages"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .any(|message| message["message"] == "durable historical message"));
-
-    drop(store);
-    let restored = SessionStore::with_persistence(&ledger, 10, 20);
-    assert_eq!(
-        restored.hot_payload_entry_count_for_test(&closed.session_id),
-        None,
-        "restored closed session must not keep the parsed event/message graph"
-    );
-    assert!(restored
-        .cold_payload_bytes_for_test(&closed.session_id)
-        .is_some_and(|bytes| bytes > 0));
-    assert!(restored
-        .hot_payload_entry_count_for_test(&active.session_id)
-        .is_some());
-    assert_eq!(
-        restored.cold_payload_bytes_for_test(&active.session_id),
-        None
-    );
-
-    let after_summary = restored.summary(&closed.session_id, Some(100)).unwrap();
-    let after_messages = restored
-        .list_messages(&closed.session_id, ListSessionMessagesFilter::default())
-        .unwrap();
-    assert_eq!(
-        serde_json::to_value(&after_summary).unwrap(),
-        serde_json::to_value(&before_summary).unwrap()
-    );
-    assert_eq!(after_messages.len(), before_messages.len());
-    assert_eq!(after_messages[0].message, before_messages[0].message);
-    assert_eq!(after_summary.lifecycle, SessionLifecycle::Closed);
-    assert_eq!(
-        restored.hot_payload_entry_count_for_test(&closed.session_id),
-        None,
-        "restart query must materialize only a temporary target record"
-    );
-    assert_eq!(
-        restored
-            .summary(&active.session_id, Some(10))
-            .unwrap()
-            .lifecycle,
-        SessionLifecycle::Active
-    );
-}
-
-#[test]
-fn cold_session_query_touch_preserves_lru_capacity_order() {
-    let store = SessionStore::new(2, 10);
-    let first = store.start_session(None, Some("cold survivor".to_string()));
-    store.close_session(&first.session_id).unwrap();
-    let second = store.start_session(None, Some("old active".to_string()));
-
-    assert!(store
-        .cold_payload_bytes_for_test(&first.session_id)
-        .is_some());
-    store.summary(&first.session_id, Some(10)).unwrap();
-    let third = store.start_session(None, Some("new active".to_string()));
-
-    assert!(store.contains_session(&first.session_id));
-    assert!(!store.contains_session(&second.session_id));
-    assert!(store.contains_session(&third.session_id));
-    assert_eq!(
-        store.lifecycle_state(&first.session_id),
-        Some(SessionLifecycle::Closed)
-    );
-    assert!(store
-        .cold_payload_bytes_for_test(&first.session_id)
-        .is_some());
-    assert!(store
-        .hot_payload_entry_count_for_test(&third.session_id)
-        .is_some());
-}
-
-#[test]
-fn closed_lifecycle_persists_round_trip() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger = tmp.path().join("sessions.json");
-    let store = persistent_store(ledger.clone());
-    let session = store.start_session(Some("proj".to_string()), Some("close persist".to_string()));
-    store.close_session(&session.session_id).unwrap();
-
-    store.flush_persistence();
-    let raw = std::fs::read_to_string(&ledger).unwrap();
-    let ledger_value: Value = serde_json::from_str(&raw).unwrap();
-    assert_eq!(ledger_value["version"], SESSION_LEDGER_VERSION);
-    assert_eq!(ledger_value["sessions"][0]["lifecycle"], "closed");
-
-    let restored = flush_and_restore(&store, ledger);
-    let summary = restored.summary(&session.session_id, Some(20)).unwrap();
-    assert_eq!(summary.lifecycle, SessionLifecycle::Closed);
-    assert!(summary
-        .events
-        .iter()
-        .any(|event| event.kind == "session_closed"));
-}
-
-#[test]
-fn closed_session_denies_mutation_tools_allows_query() {
-    let store = SessionStore::default();
-    let session = store.start_session(None, Some("closed query".to_string()));
-    store.close_session(&session.session_id).unwrap();
-
-    // Write / shell blocked.
-    let write_denial = store
-        .lifecycle_denial(
-            &session.session_id,
-            "write_project_file",
-            crate::tool_runtime::sessions::session_tool_contract("write_project_file"),
-        )
-        .expect("write denied on closed");
-    assert_eq!(write_denial.lifecycle, SessionLifecycle::Closed);
-    let shell_denial = store
-        .lifecycle_denial(
-            &session.session_id,
-            "run_shell",
-            crate::tool_runtime::sessions::session_tool_contract("run_shell"),
-        )
-        .expect("shell denied on closed");
-    assert_eq!(shell_denial.lifecycle, SessionLifecycle::Closed);
-    assert!(store
-        .lifecycle_denial(
-            &session.session_id,
-            "post_session_message",
-            crate::tool_runtime::sessions::session_tool_contract("post_session_message")
-        )
-        .is_some());
-    assert!(store
-        .lifecycle_denial(
-            &session.session_id,
-            "workspace_checkpoint_create",
-            crate::tool_runtime::sessions::session_tool_contract("workspace_checkpoint_create")
-        )
-        .is_some());
-
-    // Query / pure read still allowed; close remains idempotent path.
-    assert!(store
-        .lifecycle_denial(
-            &session.session_id,
-            "read_file",
-            crate::tool_runtime::sessions::session_tool_contract("read_file")
-        )
-        .is_none());
-    assert!(store
-        .lifecycle_denial(
-            &session.session_id,
-            "session_summary",
-            crate::tool_runtime::sessions::session_tool_contract("session_summary")
-        )
-        .is_none());
-    assert!(store
-        .lifecycle_denial(
-            &session.session_id,
-            "close_session",
-            crate::tool_runtime::sessions::session_tool_contract("close_session")
-        )
-        .is_none());
-
-    // Message board mutations fail with SessionClosed; list still works.
-    let post = store.post_message(PostSessionMessageInput {
-        session_id: session.session_id.clone(),
-        kind: SessionMessageKind::Note,
-        message: "after close".to_string(),
-        tags: Vec::new(),
-        reply_to: None,
-        priority: SessionMessagePriority::Normal,
-    });
-    assert!(matches!(
-        post,
-        Err(SessionMessageError::SessionClosed {
-            lifecycle: SessionLifecycle::Closed
-        })
-    ));
-    let listed = store
-        .list_messages(
-            &session.session_id,
-            ListSessionMessagesFilter {
-                kind: None,
-                status: None,
-                message_id: None,
-                reply_to: None,
-                limit: None,
-            },
-        )
-        .unwrap();
-    assert!(listed.is_empty());
-
-    let summary = store.summary(&session.session_id, Some(10)).unwrap();
-    assert_eq!(summary.lifecycle, SessionLifecycle::Closed);
-    assert_eq!(summary.session_id, session.session_id);
-}
-
-#[test]
-fn repeated_close_is_idempotent_without_duplicate_events() {
-    let store = SessionStore::default();
-    let session = store.start_session(None, Some("idempotent".to_string()));
-    let first = store.close_session(&session.session_id).unwrap();
-    assert!(!first.already_closed);
-    let second = store.close_session(&session.session_id).unwrap();
-    assert!(second.already_closed);
-    assert_eq!(second.summary.lifecycle, SessionLifecycle::Closed);
-
-    let summary = store.summary(&session.session_id, Some(50)).unwrap();
-    let closed_count = summary
-        .events
-        .iter()
-        .filter(|event| event.kind == "session_closed")
-        .count();
-    assert_eq!(
-        closed_count, 1,
-        "repeat close must not append another event"
-    );
-}
-
-#[test]
-fn unknown_session_close_fails_without_create() {
-    let store = SessionStore::default();
-    let missing = "wc_sess_missingclose01";
-    let err = store.close_session(missing).unwrap_err();
-    assert_eq!(err, SessionCloseError::UnknownSession);
-    assert!(!store.contains_session(missing));
-    assert!(store.summary(missing, None).is_none());
-}
-
-#[test]
-fn eviction_does_not_produce_closed_lifecycle() {
-    // Capacity eviction removes the record; it is not a Closed transition.
-    let store = SessionStore::new(1, 10);
-    let first = store.start_session(None, Some("evict me".to_string()));
-    let _second = store.start_session(None, Some("survivor".to_string()));
-    assert!(!store.contains_session(&first.session_id));
-    assert!(store.summary(&first.session_id, None).is_none());
-    // Evicted id is unknown, not Closed — close must not invent a session.
-    assert_eq!(
-        store.close_session(&first.session_id).unwrap_err(),
-        SessionCloseError::UnknownSession
-    );
-    assert!(!store.contains_session(&first.session_id));
-}
-
-#[test]
-fn closed_session_does_not_reopen() {
-    let store = SessionStore::default();
-    let session = store.start_session(None, Some("no reopen".to_string()));
-    store.close_session(&session.session_id).unwrap();
-    // Only path that could "reopen" would be inventing Active; close stays Closed.
-    let again = store.close_session(&session.session_id).unwrap();
-    assert!(again.already_closed);
-    assert_eq!(again.summary.lifecycle, SessionLifecycle::Closed);
-    assert!(!again.summary.lifecycle.allows_mutation());
-}
-
-#[test]
-fn session_context_revision_exact_missing_stale_invalid_future_and_failure_semantics() {
-    let store = SessionStore::new(10, 100);
-    let session = store.start_session(Some("proj".to_string()), Some("continuity".to_string()));
-
-    let first = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "apply_text_edits",
-        SessionContextRevisionAck::Unacknowledged,
-        true,
-        json!({"state_changed": true}),
-    );
-    assert_eq!(first.pre_call_context_revision, 0);
-    assert_eq!(first.pre_response_context_revision, 0);
-    assert_eq!(first.context_revision, 1);
-    assert!(first.checkpoint_advanced);
-
-    let second = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "work_on_project",
-        SessionContextRevisionAck::Revision(1),
-        true,
-        json!({"session_id": session.session_id}),
-    );
-    assert_eq!(second.pre_call_context_revision, 1);
-    assert_eq!(second.pre_response_context_revision, 1);
-    assert_eq!(second.context_revision, 1);
-    assert!(!second.checkpoint_advanced);
-    assert!(second.recovery_events.is_empty());
-    assert!(!second.history_lost);
-
-    let missing = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "apply_text_edits",
-        SessionContextRevisionAck::Unacknowledged,
-        true,
-        json!({"state_changed": true}),
-    );
-    assert_eq!(missing.context_revision, 2);
-    assert!(
-        missing.recovery_events.is_empty(),
-        "unknown caller state must not be treated as an acknowledged revision-zero delta"
-    );
-    assert!(!missing.history_lost);
-
-    let stale = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "work_on_project",
-        SessionContextRevisionAck::Revision(1),
-        true,
-        json!({"session_id": session.session_id}),
-    );
-    assert_eq!(stale.pre_call_context_revision, 2);
-    assert_eq!(stale.pre_response_context_revision, 2);
-    assert_eq!(stale.context_revision, 2);
-    assert!(!stale.checkpoint_advanced);
-    assert_eq!(
-        stale
-            .recovery_events
-            .iter()
-            .filter_map(|event| event.context_revision)
-            .collect::<Vec<_>>(),
-        vec![2]
-    );
-
-    let future = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "work_on_project",
-        SessionContextRevisionAck::Revision(999),
-        true,
-        json!({"session_id": session.session_id}),
-    );
-    assert_eq!(future.context_revision, 2);
-    assert_eq!(future.pre_call_context_revision, 2);
-    assert!(future.recovery_events.is_empty());
-    assert!(!future.history_lost);
-
-    let invalid = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "work_on_project",
-        SessionContextRevisionAck::Invalid,
-        true,
-        json!({"session_id": session.session_id}),
-    );
-    assert_eq!(invalid.context_revision, 2);
-    assert_eq!(invalid.pre_call_context_revision, 2);
-    assert!(invalid.recovery_events.is_empty());
-    assert!(!invalid.history_lost);
-
-    let failed = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "run_process",
-        SessionContextRevisionAck::Revision(2),
-        false,
-        json!({"failure_kind": "process_failed", "command_started": true, "command_completed": true}),
-    );
-    assert_eq!(failed.pre_call_context_revision, 2);
-    assert_eq!(failed.pre_response_context_revision, 2);
-    assert_eq!(failed.context_revision, 3);
-    assert!(failed.checkpoint_advanced);
-    assert_eq!(store.context_revision(&session.session_id), Some(3));
-}
-
-#[test]
-fn non_capable_model_facing_result_advances_cross_surface_recovery_watermark() {
-    let store = SessionStore::new(10, 100);
-    let session = store.start_session(Some("proj".to_string()), Some("cross surface".to_string()));
-    let seed = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "apply_text_edits",
-        SessionContextRevisionAck::Unacknowledged,
-        true,
-        json!({"state_changed": true}),
-    );
-    assert_eq!(seed.context_revision, 1);
-
-    let non_capable = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "workspace_checkpoint_create",
-        SessionContextRevisionAck::Unsupported,
-        true,
-        json!({
-            "checkpoint_id": "wc_ckpt_cross_surface",
-            "branch": "feature",
-            "status_summary": {"modified": 3},
-            "state_changed": true
-        }),
-    );
-    assert_eq!(non_capable.pre_call_context_revision, 1);
-    assert_eq!(non_capable.context_revision, 2);
-    assert!(non_capable.recovery_events.is_empty());
-    assert!(!non_capable.history_lost);
-    assert_eq!(store.context_revision(&session.session_id), Some(2));
-
-    let resumed = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "work_on_project",
-        SessionContextRevisionAck::Revision(1),
-        true,
-        json!({"session_id": session.session_id}),
-    );
-    assert_eq!(resumed.pre_call_context_revision, 2);
-    assert_eq!(resumed.pre_response_context_revision, 2);
-    assert_eq!(resumed.context_revision, 2);
-    assert!(!resumed.checkpoint_advanced);
-    assert_eq!(
-        resumed
-            .recovery_events
-            .iter()
-            .filter_map(|event| event.context_revision)
-            .collect::<Vec<_>>(),
-        vec![2]
-    );
-    let checkpoint = &resumed.recovery_events[0];
-    assert_eq!(checkpoint.tool_name, "workspace_checkpoint_create");
-    let consequence = checkpoint.context_result_summary.as_ref().unwrap();
-    assert_eq!(consequence["checkpoint_id"], "wc_ckpt_cross_surface");
-    assert_eq!(consequence["status_summary"]["modified"], 3);
-}
-
-#[test]
-fn session_context_revision_retention_reports_history_loss_without_counter_regression() {
-    let store = SessionStore::new(10, 4);
-    let session = store.start_session(Some("proj".to_string()), Some("retention".to_string()));
-    let mut latest = 0;
-    for _ in 0..6 {
-        let recorded = record_model_facing_result(
-            &store,
-            &session.session_id,
-            "apply_text_edits",
-            SessionContextRevisionAck::Revision(latest),
-            true,
-            json!({"state_changed": true}),
-        );
-        latest = recorded.context_revision;
-    }
-    assert_eq!(latest, 6);
-    assert_eq!(store.context_revision(&session.session_id), Some(6));
-
-    let stale = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "work_on_project",
-        SessionContextRevisionAck::Revision(1),
-        true,
-        json!({"session_id": session.session_id}),
-    );
-    assert_eq!(stale.pre_call_context_revision, 6);
-    assert_eq!(stale.pre_response_context_revision, 6);
-    assert_eq!(stale.context_revision, 6);
-    assert!(!stale.checkpoint_advanced);
-    assert!(stale.history_lost);
-    assert!(stale.recovery_events.len() < 5);
 }
 
 #[test]
@@ -5366,53 +3110,6 @@ fn history_lost_detects_evicted_checkpoint_after_raw_event_churn() {
 }
 
 #[test]
-fn generic_session_events_do_not_advance_model_context_revision() {
-    let store = SessionStore::new(10, 100);
-    let session = store.start_session(Some("proj".to_string()), Some("background".to_string()));
-    let first = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "apply_text_edits",
-        SessionContextRevisionAck::Unacknowledged,
-        true,
-        json!({"state_changed": true}),
-    );
-    assert_eq!(first.context_revision, 1);
-
-    let generic_start = store.record_tool_call_started(
-        Some(&session.session_id),
-        SessionTransport::Api,
-        "cargo_test",
-        &json!({"project": "proj"}),
-        crate::tool_runtime::sessions::session_tool_contract("cargo_test"),
-    );
-    assert!(store
-        .record_tool_call_finished(
-            generic_start,
-            true,
-            &json!({"tests_run": 1, "passed": true}),
-            None,
-            None,
-        )
-        .is_some());
-    assert_eq!(store.context_revision(&session.session_id), Some(1));
-
-    let exact = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "work_on_project",
-        SessionContextRevisionAck::Revision(1),
-        true,
-        json!({"session_id": session.session_id}),
-    );
-    assert_eq!(exact.pre_call_context_revision, 1);
-    assert_eq!(exact.pre_response_context_revision, 1);
-    assert_eq!(exact.context_revision, 1);
-    assert!(!exact.checkpoint_advanced);
-    assert!(exact.recovery_events.is_empty());
-}
-
-#[test]
 fn simultaneous_model_facing_results_allocate_unique_ordered_revisions() {
     let store = SessionStore::new(10, 100);
     let session = store.start_session(Some("proj".to_string()), Some("concurrent".to_string()));
@@ -5589,158 +3286,6 @@ fn simultaneous_model_facing_results_allocate_unique_ordered_revisions() {
         future_response.output["session_continuity"]["status"],
         "invalid"
     );
-}
-
-#[test]
-fn simultaneous_no_checkpoint_results_leave_revision_unchanged() {
-    let store = SessionStore::new(10, 100);
-    let session = store.start_session(
-        Some("proj".to_string()),
-        Some("read concurrency".to_string()),
-    );
-    let seed = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "apply_text_edits",
-        SessionContextRevisionAck::Unacknowledged,
-        true,
-        json!({"state_changed": true}),
-    );
-    assert_eq!(seed.context_revision, 1);
-
-    let make_start = |tool_name: &str| {
-        store
-            .record_tool_call_started_with_metadata(
-                Some(&session.session_id),
-                SessionTransport::Mcp,
-                tool_name,
-                &json!({"project": "proj"}),
-                Some("proj".to_string()),
-                ToolCallRecorderMetadata {
-                    ack_session_context_revision: SessionContextRevisionAck::Revision(1),
-                    ..Default::default()
-                },
-                crate::tool_runtime::sessions::session_tool_contract(tool_name),
-            )
-            .unwrap()
-    };
-    let read_start = make_start("read_file");
-    let search_start = make_start("search_project_texts");
-    assert_eq!(read_start.pre_call_context_revision, 1);
-    assert_eq!(search_start.pre_call_context_revision, 1);
-
-    let a_store = store.clone();
-    let b_store = store.clone();
-    let read = std::thread::spawn(move || {
-        a_store
-            .record_model_facing_tool_call_finished(
-                Some(read_start),
-                true,
-                &json!({"content": "read"}),
-                None,
-                None,
-            )
-            .unwrap()
-    });
-    let search = std::thread::spawn(move || {
-        b_store
-            .record_model_facing_tool_call_finished(
-                Some(search_start),
-                true,
-                &json!({"matches": []}),
-                None,
-                None,
-            )
-            .unwrap()
-    });
-    for recorded in [read.join().unwrap(), search.join().unwrap()] {
-        assert_eq!(recorded.context_revision, 1);
-        assert_eq!(recorded.pre_response_context_revision, 1);
-        assert!(!recorded.checkpoint_advanced);
-        assert_eq!(
-            recorded.ack_session_context_revision,
-            SessionContextRevisionAck::Revision(1)
-        );
-    }
-    assert_eq!(store.context_revision(&session.session_id), Some(1));
-}
-
-#[test]
-fn concurrent_checkpoint_and_raw_result_advance_once() {
-    let store = SessionStore::new(10, 100);
-    let session = store.start_session(
-        Some("proj".to_string()),
-        Some("mixed concurrency".to_string()),
-    );
-    let seed = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "apply_text_edits",
-        SessionContextRevisionAck::Unacknowledged,
-        true,
-        json!({"state_changed": true}),
-    );
-    assert_eq!(seed.context_revision, 1);
-
-    let read_start = store
-        .record_tool_call_started_with_metadata(
-            Some(&session.session_id),
-            SessionTransport::Mcp,
-            "read_file",
-            &json!({"project": "proj"}),
-            Some("proj".to_string()),
-            ToolCallRecorderMetadata::default(),
-            crate::tool_runtime::sessions::session_tool_contract("read_file"),
-        )
-        .unwrap();
-    let checkpoint_start = store
-        .record_tool_call_started_with_metadata(
-            Some(&session.session_id),
-            SessionTransport::Mcp,
-            "run_process",
-            &json!({"project": "proj"}),
-            Some("proj".to_string()),
-            ToolCallRecorderMetadata {
-                ack_session_context_revision: SessionContextRevisionAck::Revision(1),
-                ..Default::default()
-            },
-            crate::tool_runtime::sessions::session_tool_contract("run_process"),
-        )
-        .unwrap();
-    assert_eq!(read_start.pre_call_context_revision, 1);
-    assert_eq!(checkpoint_start.pre_call_context_revision, 1);
-
-    let a_store = store.clone();
-    let b_store = store.clone();
-    let read = std::thread::spawn(move || {
-        a_store
-            .record_model_facing_tool_call_finished(
-                Some(read_start),
-                true,
-                &json!({"content": "read"}),
-                None,
-                None,
-            )
-            .unwrap()
-    });
-    let checkpoint = std::thread::spawn(move || {
-        b_store
-            .record_model_facing_tool_call_finished(
-                Some(checkpoint_start),
-                true,
-                &json!({"command_started": true, "command_completed": true, "exit_code": 0}),
-                None,
-                None,
-            )
-            .unwrap()
-    });
-    let read = read.join().unwrap();
-    let checkpoint = checkpoint.join().unwrap();
-    assert!(!read.checkpoint_advanced);
-    assert!(matches!(read.context_revision, 1 | 2));
-    assert!(checkpoint.checkpoint_advanced);
-    assert_eq!(checkpoint.context_revision, 2);
-    assert_eq!(store.context_revision(&session.session_id), Some(2));
 }
 
 #[test]
@@ -6060,171 +3605,6 @@ async fn history_lost_session_context_recovery_adds_current_handoff() {
         .add_session_history_recovery(&mut response, &behind, None)
         .await;
     assert!(response.output["session_recovery"]["current_handoff"].is_object());
-}
-
-#[test]
-fn session_context_revision_survives_persistence_restart() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger = tmp.path().join("sessions.json");
-    let store = SessionStore::with_persistence(&ledger, 10, 100);
-    let session = store.start_session(Some("proj".to_string()), Some("restart".to_string()));
-    let first = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "apply_text_edits",
-        SessionContextRevisionAck::Unacknowledged,
-        true,
-        json!({"state_changed": true}),
-    );
-    assert_eq!(first.context_revision, 1);
-    let raw = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "read_file",
-        SessionContextRevisionAck::Revision(1),
-        true,
-        json!({"content": "persisted raw observation"}),
-    );
-    assert_eq!(raw.context_revision, 1);
-    assert!(!raw.checkpoint_advanced);
-    assert_eq!(store.context_revision(&session.session_id), Some(1));
-    store.flush_persistence();
-    drop(store);
-
-    let restored = SessionStore::with_persistence(&ledger, 10, 100);
-    assert_eq!(restored.context_revision(&session.session_id), Some(1));
-    let second = record_model_facing_result(
-        &restored,
-        &session.session_id,
-        "run_process",
-        SessionContextRevisionAck::Revision(1),
-        true,
-        json!({"command_started": true, "command_completed": true, "exit_code": 0}),
-    );
-    assert_eq!(second.pre_call_context_revision, 1);
-    assert_eq!(second.pre_response_context_revision, 1);
-    assert_eq!(second.context_revision, 2);
-    assert!(second.checkpoint_advanced);
-    assert!(second.recovery_events.is_empty());
-}
-
-#[test]
-fn existing_persisted_context_watermark_is_not_renumbered() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger = tmp.path().join("sessions.json");
-    let store = SessionStore::with_persistence(&ledger, 10, 100);
-    let session = store.start_session(Some("proj".to_string()), Some("high watermark".to_string()));
-    let first = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "apply_text_edits",
-        SessionContextRevisionAck::Unacknowledged,
-        true,
-        json!({"state_changed": true}),
-    );
-    assert_eq!(first.context_revision, 1);
-    store.flush_persistence();
-    drop(store);
-
-    let mut persisted: Value = serde_json::from_slice(&std::fs::read(&ledger).unwrap()).unwrap();
-    persisted["sessions"][0]["context_revision"] = json!(847);
-    std::fs::write(&ledger, serde_json::to_vec(&persisted).unwrap()).unwrap();
-
-    let restored = SessionStore::with_persistence(&ledger, 10, 100);
-    assert_eq!(restored.context_revision(&session.session_id), Some(847));
-    let read = record_model_facing_result(
-        &restored,
-        &session.session_id,
-        "read_file",
-        SessionContextRevisionAck::Revision(847),
-        true,
-        json!({"content": "still 847"}),
-    );
-    assert_eq!(read.context_revision, 847);
-    assert!(!read.checkpoint_advanced);
-    let next = record_model_facing_result(
-        &restored,
-        &session.session_id,
-        "run_process",
-        SessionContextRevisionAck::Revision(847),
-        true,
-        json!({"command_started": true, "command_completed": true, "exit_code": 0}),
-    );
-    assert_eq!(next.context_revision, 848);
-}
-
-#[test]
-fn session_context_revision_restore_revalidates_bounded_context_result_summary() {
-    let tmp = tempfile::tempdir().unwrap();
-    let ledger = tmp.path().join("sessions.json");
-    let store = SessionStore::with_persistence(&ledger, 10, 100);
-    let session = store.start_session(
-        Some("proj".to_string()),
-        Some("context sanitize".to_string()),
-    );
-    let recorded = record_model_facing_result(
-        &store,
-        &session.session_id,
-        "workspace_checkpoint_create",
-        SessionContextRevisionAck::Unacknowledged,
-        true,
-        json!({
-            "checkpoint_id": "wc_ckpt_demo",
-            "branch": "b".repeat(500),
-            "status_summary": {
-                "modified": 21,
-                "token": "wc_pat_must_not_survive"
-            }
-        }),
-    );
-    assert_eq!(recorded.context_revision, 1);
-    let live = store.summary(&session.session_id, Some(20)).unwrap();
-    let live_summary = live
-        .events
-        .iter()
-        .find(|event| event.context_revision == Some(1))
-        .and_then(|event| event.context_result_summary.as_ref())
-        .unwrap();
-    assert_eq!(live_summary["status_summary"]["modified"], 21);
-    assert_eq!(live_summary["status_summary"]["token"], "[redacted]");
-    let live_branch = live_summary["branch"].as_str().unwrap();
-    assert!(live_branch.chars().count() <= 123);
-    assert!(live_branch.ends_with("..."));
-
-    store.flush_persistence();
-    drop(store);
-    let mut persisted: Value = serde_json::from_slice(&std::fs::read(&ledger).unwrap()).unwrap();
-    let events = persisted["sessions"][0]["events"].as_array_mut().unwrap();
-    let finished = events
-        .iter_mut()
-        .find(|event| event["context_revision"] == 1)
-        .unwrap();
-    finished["context_result_summary"] = json!({
-        "checkpoint_id": "wc_ckpt_demo",
-        "branch": "x".repeat(500),
-        "status_summary": {
-            "modified": 21,
-            "token": "wc_pat_corrupt_ledger_secret"
-        },
-        "arbitrary_untrusted_body": "must not survive restore"
-    });
-    std::fs::write(&ledger, serde_json::to_vec(&persisted).unwrap()).unwrap();
-
-    let restored = SessionStore::with_persistence(&ledger, 10, 100);
-    assert_eq!(restored.context_revision(&session.session_id), Some(1));
-    let restored_summary = restored.summary(&session.session_id, Some(20)).unwrap();
-    let context = restored_summary
-        .events
-        .iter()
-        .find(|event| event.context_revision == Some(1))
-        .and_then(|event| event.context_result_summary.as_ref())
-        .unwrap();
-    assert!(context.get("arbitrary_untrusted_body").is_none());
-    assert_eq!(context["status_summary"]["modified"], 21);
-    assert_eq!(context["status_summary"]["token"], "[redacted]");
-    let restored_branch = context["branch"].as_str().unwrap();
-    assert!(restored_branch.chars().count() <= 123);
-    assert!(restored_branch.ends_with("..."));
 }
 
 #[test]
