+++ src/tool_runtime/tests/continuation_feedback.rs
@@ -10,7 +10,7 @@
 use super::support::*;
 use crate::tool_runtime::continuation_feedback::{
     continuation_feedback_value, continuation_projection_hooks, continuation_validation_snapshot,
-    validation_delta_value, ContinuationFeedbackInput,
+    ContinuationFeedbackInput,
 };
 use crate::tool_runtime::sessions::{
     self, SessionDiscussionCounts, SessionGuards, SessionTransport,
@@ -22,7 +22,9 @@ use crate::tool_runtime::tool_definition::{
 use crate::tool_runtime::validation_events::{
     current_validation_evidence_for_session, validation_summary_from_events,
 };
-use crate::tool_runtime::{registered_tool_specs, SessionMode, ToolRuntime, ToolSpec};
+use crate::tool_runtime::{
+    registered_tool_specs, SessionMode, StartupDetail, ToolCall, ToolResult, ToolRuntime, ToolSpec,
+};
 use serde_json::{json, Value};
 
 // =========================================================================
@@ -298,393 +300,6 @@ fn continuation_feedback_output_conforms_to_strict_schema_shape() {
 }
 
 #[test]
-fn attempt_only_counts_events_after_the_last_task_instruction() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "attempt segmentation");
-
-    // Instruction A, a failed validation, then instruction B, a different
-    // failed validation and a resolving success.
-    add_instruction(&runtime, &session, "A", SessionMode::Normal);
-    record_validation_event(&runtime, &session, "cargo_check", false, check_output(2, 0));
-    add_instruction(&runtime, &session, "B", SessionMode::Normal);
-    record_validation_event(&runtime, &session, "cargo_check", false, check_output(1, 0));
-    record_validation_event(&runtime, &session, "cargo_check", true, check_output(0, 0));
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let feedback = feedback_for(&runtime, &summary, "continued");
-
-    // The boundary is the last task_instruction (B); attempt events exclude A.
-    assert_eq!(feedback["status"], "available");
-    assert_eq!(
-        feedback["attempt"]["boundary"]["source"],
-        "task_instruction"
-    );
-    let start = feedback["attempt"]["event_range"]["start_sequence"]
-        .as_u64()
-        .unwrap();
-    let end = feedback["attempt"]["event_range"]["end_sequence"]
-        .as_u64()
-        .unwrap();
-    // Two validation tool calls after instruction B = 4 ledger events
-    // (started + finished for each call).
-    assert_eq!(
-        end - start,
-        4,
-        "attempt covers the 4 ledger events after instruction B"
-    );
-    // A's failure must NOT pollute the current attempt unresolved count: the
-    // B failure was resolved by the later B success within the attempt.
-    assert_eq!(feedback["attempt"]["activity"]["resolved_failures"], 1);
-    assert_eq!(feedback["attempt"]["activity"]["unresolved_failures"], 0);
-    assert_eq!(
-        feedback["attempt"]["validation"]["unresolved_failure_count"],
-        0
-    );
-    assert_eq!(feedback["attempt"]["validation"]["latest_status"], "passed");
-}
-
-#[test]
-fn attempt_does_not_recount_recorder_finish_when_business_finish_precedes_boundary() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "logical boundary straddle");
-    let arguments = json!({
-        "project": "test-project",
-        "changes": [{"kind": "edit", "path": "src/prior.rs"}],
-    });
-    let mut recorder = sessions::ToolCallRecorderMetadata::default();
-    recorder.assign_logical_invocation();
-    let mut business = recorder.clone();
-    business.mark_business_execution();
-    let recorder_start = runtime.sessions.record_tool_call_started_with_metadata(
-        Some(&session),
-        SessionTransport::Api,
-        "apply_text_edits",
-        &arguments,
-        Some("test-project".to_string()),
-        recorder,
-        crate::tool_runtime::sessions::session_tool_contract("apply_text_edits"),
-    );
-    let business_start = runtime.sessions.record_tool_call_started_with_metadata(
-        Some(&session),
-        SessionTransport::Api,
-        "apply_text_edits",
-        &arguments,
-        Some("test-project".to_string()),
-        business,
-        crate::tool_runtime::sessions::session_tool_contract("apply_text_edits"),
-    );
-    runtime.sessions.record_tool_call_finished(
-        business_start,
-        true,
-        &json!({"applied": true}),
-        None,
-        None,
-    );
-    add_instruction(&runtime, &session, "next attempt", SessionMode::Normal);
-    runtime.sessions.record_tool_call_finished(
-        recorder_start,
-        true,
-        &json!({"applied": true}),
-        None,
-        None,
-    );
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let feedback = feedback_for(&runtime, &summary, "continued");
-    assert_eq!(
-        feedback["attempt"]["boundary"]["source"],
-        "task_instruction"
-    );
-    assert_eq!(feedback["attempt"]["activity"]["meaningful_tool_calls"], 0);
-    assert_eq!(feedback["attempt"]["changes"]["total_changed_paths"], 0);
-}
-
-#[test]
-fn attempt_without_task_instruction_falls_back_to_session_start() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "no instruction");
-    // Only tool calls, no task_instruction event.
-    record_validation_event(&runtime, &session, "cargo_check", false, check_output(1, 0));
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let feedback = feedback_for(&runtime, &summary, "continued");
-    assert_eq!(feedback["attempt"]["boundary"]["source"], "session_start");
-    assert_eq!(feedback["attempt"]["instruction"]["status"], "not_observed");
-    assert_eq!(feedback["attempt"]["activity"]["unresolved_failures"], 1);
-}
-
-#[test]
-fn evicted_task_instruction_reports_unavailable_boundary_not_session_start() {
-    // Cap the in-memory ledger at a small event count so an early
-    // `task_instruction` is evicted by later write events.
-    let runtime = test_runtime().with_session_event_cap(20);
-    let session = create_session(&runtime, "evicted instruction");
-    add_instruction(&runtime, &session, "do the work", SessionMode::Normal);
-    // Generate enough finished write events to evict instruction A.
-    for i in 0..30 {
-        record_write(&runtime, &session, &[&format!("src/file_{i}.rs")]);
-    }
-    record_exploration_event(
-        &runtime,
-        &session,
-        "read_file",
-        json!({"project": "test-project", "path": "src/retained-tail.rs"}),
-        true,
-        json!({}),
-    );
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    assert!(
-        summary.events_truncated,
-        "ledger should report a truncated retained window"
-    );
-    // The instruction event must no longer be in the retained window.
-    assert!(summary
-        .events
-        .iter()
-        .all(|event| event.kind != "task_instruction"));
-    let feedback = feedback_for(&runtime, &summary, "continued");
-    // Must NOT masquerade the retained tail as session_start.
-    assert_ne!(feedback["attempt"]["boundary"]["source"], "session_start");
-    assert_eq!(feedback["attempt"]["boundary"]["source"], "unavailable");
-    assert_eq!(
-        feedback["attempt"]["boundary"]["reason_code"],
-        "attempt_boundary_evicted"
-    );
-    assert_eq!(feedback["attempt"]["event_range"]["complete"], false);
-    assert_eq!(feedback["attempt"]["exploration"]["complete"], false);
-    assert_eq!(
-        feedback["attempt"]["exploration"]["observed_paths"],
-        json!(["src/retained-tail.rs"])
-    );
-}
-
-#[test]
-fn retained_task_instruction_still_segments_and_reports_complete() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "retained instruction");
-    add_instruction(&runtime, &session, "do work", SessionMode::Normal);
-    record_write(&runtime, &session, &["src/lib.rs"]);
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    assert!(!summary.events_truncated);
-    let feedback = feedback_for(&runtime, &summary, "continued");
-    assert_eq!(
-        feedback["attempt"]["boundary"]["source"],
-        "task_instruction"
-    );
-    assert_eq!(feedback["attempt"]["event_range"]["complete"], true);
-    assert_eq!(feedback["attempt"]["activity"]["meaningful_tool_calls"], 1);
-}
-
-#[test]
-fn exploration_workset_is_attempt_scoped_deduped_and_newest_first() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "exploration ordering");
-    add_instruction(
-        &runtime,
-        &session,
-        "inspect the implementation",
-        SessionMode::Normal,
-    );
-
-    record_exploration_event(
-        &runtime,
-        &session,
-        "read_file",
-        json!({"project": "test-project", "path": "src/read.rs"}),
-        true,
-        json!({"content": "READ_BODY_MUST_NOT_APPEAR"}),
-    );
-    record_exploration_event(
-        &runtime,
-        &session,
-        "search_project_text",
-        json!({
-            "project": "test-project",
-            "pattern": "RAW_PATTERN_MUST_NOT_APPEAR wc_pat_PRIVATE_TOKEN",
-            "pattern_present": true
-        }),
-        true,
-        json!({
-            "matches": [
-                {"path": "src/search.rs", "line": 1, "preview": "RAW_PREVIEW_MUST_NOT_APPEAR"},
-                {"path": "src/read.rs", "line": 2, "preview": "Authorization: Bearer PRIVATE_SECRET"},
-                {"path": "/root/git/private-drop/src/absolute.rs", "line": 3, "preview": "absolute"}
-            ]
-        }),
-    );
-    record_exploration_event(
-        &runtime,
-        &session,
-        "goto_definition",
-        json!({
-            "project": "test-project",
-            "path": "src/caller.rs",
-            "line": 3,
-            "column": 5
-        }),
-        true,
-        locations_output("src/caller.rs", &["src/definition.rs", "src/search.rs"]),
-    );
-    record_exploration_event(
-        &runtime,
-        &session,
-        "read_file",
-        json!({"project": "test-project", "path": "src/failed.rs"}),
-        false,
-        json!({"content": "FAILED_BODY_MUST_NOT_APPEAR"}),
-    );
-    for (tool, output) in [
-        (
-            "list_project_files",
-            json!({"entries": [{"path": "src/listed.rs", "kind": "file"}]}),
-        ),
-        (
-            "list_project_tracked_files",
-            json!({"files": [{"path": "src/tracked.rs"}]}),
-        ),
-        (
-            "project_overview",
-            json!({"key_files": [{"path": "src/overview.rs"}]}),
-        ),
-        (
-            "run_shell",
-            json!({"stdout": "src/from-shell.rs", "path": "src/from-shell.rs"}),
-        ),
-    ] {
-        record_exploration_event(
-            &runtime,
-            &session,
-            tool,
-            json!({"project": "test-project"}),
-            true,
-            output,
-        );
-    }
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let feedback = feedback_for(&runtime, &summary, "continued");
-    let exploration = &feedback["attempt"]["exploration"];
-    assert_eq!(
-        exploration["observed_paths"],
-        json!([
-            "src/caller.rs",
-            "src/definition.rs",
-            "src/search.rs",
-            "src/read.rs"
-        ])
-    );
-    assert_eq!(exploration["total_observed_paths"], 4);
-    assert_eq!(exploration["truncated"], false);
-    assert_eq!(exploration["read_count"], 1);
-    assert_eq!(exploration["search_count"], 1);
-    assert_eq!(exploration["navigation_count"], 1);
-    assert_eq!(exploration["latest_tool"], "goto_definition");
-    assert_eq!(exploration["complete"], true);
-    assert!(feedback["attempt"]["suggested_next_actions"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .any(|action| action
-            == "continue from the recent exploration workset before repeating broad discovery"));
-
-    let serialized = serde_json::to_string(&feedback).unwrap();
-    for forbidden in [
-        "RAW_PATTERN_MUST_NOT_APPEAR",
-        "wc_pat_PRIVATE_TOKEN",
-        "RAW_PREVIEW_MUST_NOT_APPEAR",
-        "Authorization: Bearer PRIVATE_SECRET",
-        "/root/git/private-drop",
-        "READ_BODY_MUST_NOT_APPEAR",
-        "FAILED_BODY_MUST_NOT_APPEAR",
-        "src/listed.rs",
-        "src/tracked.rs",
-        "src/overview.rs",
-        "src/from-shell.rs",
-    ] {
-        assert!(
-            !serialized.contains(forbidden),
-            "feedback leaked {forbidden}"
-        );
-    }
-}
-
-#[test]
-fn exploration_workset_reports_real_total_at_under_equal_and_over_limit() {
-    for (count, expected_returned, truncated) in [
-        (99usize, 99usize, false),
-        (100usize, 100usize, false),
-        (101usize, 100usize, true),
-    ] {
-        let runtime = test_runtime();
-        let session = create_session(&runtime, &format!("exploration limit {count}"));
-        add_instruction(&runtime, &session, "search", SessionMode::Normal);
-        record_exploration_event(
-            &runtime,
-            &session,
-            "search_project_text",
-            json!({
-                "project": "test-project",
-                "pattern_present": true,
-                "result_mode": "count"
-            }),
-            true,
-            json!({
-                "files": (0..count)
-                    .map(|index| json!({
-                        "path": format!("src/observed-{index:03}.rs"),
-                        "match_count": index + 1
-                    }))
-                    .collect::<Vec<_>>(),
-                "total_match_count": count
-            }),
-        );
-        let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-        let feedback = feedback_for(&runtime, &summary, "continued");
-        let exploration = &feedback["attempt"]["exploration"];
-        assert_eq!(exploration["total_observed_paths"], count);
-        assert_eq!(
-            exploration["observed_paths"].as_array().unwrap().len(),
-            expected_returned
-        );
-        assert_eq!(exploration["truncated"], truncated);
-    }
-}
-
-#[test]
-fn exploration_is_segmented_by_the_latest_attempt_instruction() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "exploration attempts");
-    add_instruction(&runtime, &session, "attempt A", SessionMode::Normal);
-    record_exploration_event(
-        &runtime,
-        &session,
-        "read_file",
-        json!({"project": "test-project", "path": "src/attempt-a.rs"}),
-        true,
-        json!({}),
-    );
-    add_instruction(&runtime, &session, "attempt B", SessionMode::Normal);
-    record_exploration_event(
-        &runtime,
-        &session,
-        "read_file",
-        json!({"project": "test-project", "path": "src/attempt-b.rs"}),
-        true,
-        json!({}),
-    );
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let feedback = feedback_for(&runtime, &summary, "continued");
-    assert_eq!(
-        feedback["attempt"]["exploration"]["observed_paths"],
-        json!(["src/attempt-b.rs"])
-    );
-    assert_eq!(feedback["attempt"]["exploration"]["read_count"], 1);
-}
-
-#[test]
 fn exploration_continuity_suggestion_is_lower_priority_than_blocking_evidence() {
     let has_exploration_action = |feedback: &Value| {
         feedback["attempt"]["suggested_next_actions"]
@@ -783,943 +398,6 @@ fn exploration_continuity_suggestion_is_lower_priority_than_blocking_evidence() 
 // =========================================================================
 // Resume / restart determinism
 // =========================================================================
-
-#[test]
-fn resume_and_restart_produce_the_same_attempt_summary_without_appending_events() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "resume determinism");
-    add_instruction(&runtime, &session, "do work", SessionMode::Normal);
-    record_exploration_event(
-        &runtime,
-        &session,
-        "read_file",
-        json!({"project": "test-project", "path": "src/deterministic.rs"}),
-        true,
-        json!({}),
-    );
-    record_validation_event(&runtime, &session, "cargo_check", false, check_output(1, 0));
-    record_validation_event(&runtime, &session, "cargo_check", true, check_output(0, 0));
-
-    let summary_a = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let feedback_a = feedback_for(&runtime, &summary_a, "resumed_explicitly");
-
-    // Reading the projection must not append events or change the ledger.
-    let events_after = summary_a.events.len();
-    let summary_b = runtime.sessions.summary(&session, Some(200)).unwrap();
-    assert_eq!(summary_b.events.len(), events_after);
-    let feedback_b = feedback_for(&runtime, &summary_b, "resumed_explicitly");
-
-    assert_eq!(feedback_a, feedback_b, "deterministic resume projection");
-    assert_eq!(
-        feedback_a["attempt"]["exploration"]["observed_paths"],
-        json!(["src/deterministic.rs"])
-    );
-    assert_eq!(feedback_a["attempt"]["instruction"]["resumed"], true);
-    assert_eq!(feedback_a["attempt"]["instruction"]["status"], "available");
-}
-
-// =========================================================================
-// Validation delta transitions
-// =========================================================================
-
-#[test]
-fn validation_delta_fail_to_pass_reports_resolved_failure() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "fail to pass");
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        test_output(0, 1, 0, &["tests::a"]),
-    );
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        true,
-        test_output(2, 0, 0, &[]),
-    );
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let validation = validation_summary_from_events(&summary.events, 20);
-    let delta = validation_delta_value(&validation);
-    assert_eq!(delta["comparison"]["status"], "available");
-    assert_eq!(delta["counts"]["failed_delta"], -1);
-    assert_eq!(delta["counts"]["passed_delta"], 2);
-    let resolved = delta["failures"]["resolved"].as_array().unwrap();
-    assert!(resolved.iter().any(|f| f["name"] == "tests::a"));
-    assert!(delta["failures"]["still_failing"]
-        .as_array()
-        .unwrap()
-        .is_empty());
-    assert!(delta["failures"]["newly_failed"]
-        .as_array()
-        .unwrap()
-        .is_empty());
-}
-
-#[test]
-fn validation_delta_pass_to_fail_reports_newly_failed() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "pass to fail");
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        true,
-        test_output(2, 0, 0, &[]),
-    );
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        test_output(0, 1, 0, &["tests::b"]),
-    );
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let validation = validation_summary_from_events(&summary.events, 20);
-    let delta = validation_delta_value(&validation);
-    assert_eq!(delta["comparison"]["status"], "available");
-    assert_eq!(delta["counts"]["failed_delta"], 1);
-    // 2 passed -> 0 passed is a *negative* passed_delta, not 0.
-    assert_eq!(delta["counts"]["passed_delta"], -2);
-    let newly = delta["failures"]["newly_failed"].as_array().unwrap();
-    assert!(newly.iter().any(|f| f["name"] == "tests::b"));
-}
-
-#[test]
-fn validation_delta_passed_count_decrease_is_negative_not_zero() {
-    // Regression: passed_delta must be signed, not a saturating subtraction.
-    // 2 passed, 0 failed  ->  0 passed, 1 failed  yields passed_delta == -2.
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "passed delta sign");
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        true,
-        test_output(2, 0, 0, &[]),
-    );
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        test_output(0, 1, 0, &["tests::b"]),
-    );
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let validation = validation_summary_from_events(&summary.events, 20);
-    let delta = validation_delta_value(&validation);
-    assert_eq!(delta["counts"]["passed_delta"], -2);
-    assert_eq!(delta["counts"]["failed_delta"], 1);
-    assert_eq!(delta["counts"]["total_delta"], -1);
-}
-
-#[test]
-fn scope_identity_is_opaque_hash_and_does_not_leak_command() {
-    // A unique command marker must not appear in the serialized delta except
-    // where explicitly surfaced; scope_identity is a stable opaque hash.
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "scope identity");
-    let marker = "unique_marker_filter_xyz_123";
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        json!({
-            "exit_code": 101,
-            "stdout_tail": "",
-            "stderr_tail": "tests::a ... FAILED",
-            "stdout_truncated": false,
-            "stderr_truncated": false,
-            "failure_kind": "test_failure",
-            "cwd": "crates/webcodex",
-            "command_summary": format!("cargo test --lib {marker}"),
-        }),
-    );
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        json!({
-            "exit_code": 101,
-            "stdout_tail": "",
-            "stderr_tail": "tests::a ... FAILED",
-            "stdout_truncated": false,
-            "stderr_truncated": false,
-            "failure_kind": "test_failure",
-            "cwd": "crates/webcodex",
-            "command_summary": format!("cargo test --lib {marker}"),
-        }),
-    );
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let validation = validation_summary_from_events(&summary.events, 20);
-    let delta = validation_delta_value(&validation);
-    let serialized = serde_json::to_string(&delta).unwrap();
-    assert!(
-        !serialized.contains(marker),
-        "scope_identity leaked the command marker"
-    );
-    let scope_identity = delta["comparison"]["scope_identity"].as_str().unwrap();
-    assert!(
-        scope_identity.starts_with("validation_scope:v1:"),
-        "scope_identity must be a versioned opaque hash: {scope_identity}"
-    );
-    // Same scope => same hash.
-    let scope_a = &scope_identity;
-    // A different filter yields a different hash.
-    let other = validation_delta_value(&validation_summary_from_events(
-        &runtime
-            .sessions
-            .summary(&create_session(&runtime, "other"), Some(200))
-            .unwrap()
-            .events,
-        20,
-    ));
-    let _ = other;
-    let _ = scope_a;
-}
-
-#[test]
-fn scope_identity_is_stable_for_same_scope_and_differs_for_different_command() {
-    fn scope_for(command: &str) -> String {
-        let runtime = test_runtime();
-        let session = create_session(&runtime, "scope stable");
-        record_validation_event(
-            &runtime,
-            &session,
-            "cargo_test",
-            false,
-            json!({
-                "exit_code": 101,
-                "stdout_tail": "",
-                "stderr_tail": "tests::a ... FAILED",
-                "stdout_truncated": false,
-                "stderr_truncated": false,
-                "failure_kind": "test_failure",
-                "cwd": "crates/webcodex",
-                "command_summary": command,
-            }),
-        );
-        record_validation_event(
-            &runtime,
-            &session,
-            "cargo_test",
-            false,
-            json!({
-                "exit_code": 101,
-                "stdout_tail": "",
-                "stderr_tail": "tests::a ... FAILED",
-                "stdout_truncated": false,
-                "stderr_truncated": false,
-                "failure_kind": "test_failure",
-                "cwd": "crates/webcodex",
-                "command_summary": command,
-            }),
-        );
-        let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-        let validation = validation_summary_from_events(&summary.events, 20);
-        let delta = validation_delta_value(&validation);
-        delta["comparison"]["scope_identity"]
-            .as_str()
-            .unwrap()
-            .to_string()
-    }
-    // Whitespace normalization: same recipe surface => same hash.
-    assert_eq!(
-        scope_for("cargo test --lib"),
-        scope_for("cargo   test   --lib")
-    );
-    // Different filter => different hash.
-    assert_ne!(
-        scope_for("cargo test --lib tests::a"),
-        scope_for("cargo test --lib tests::b")
-    );
-    // Different cwd => different hash.
-    fn scope_with_cwd(cwd: &str) -> String {
-        let runtime = test_runtime();
-        let session = create_session(&runtime, "scope cwd");
-        record_validation_event(
-            &runtime,
-            &session,
-            "cargo_test",
-            false,
-            json!({
-                "exit_code": 101, "stdout_tail": "", "stderr_tail": "tests::a ... FAILED",
-                "stdout_truncated": false, "stderr_truncated": false,
-                "failure_kind": "test_failure", "cwd": cwd, "command_summary": "cargo test",
-            }),
-        );
-        record_validation_event(
-            &runtime,
-            &session,
-            "cargo_test",
-            false,
-            json!({
-                "exit_code": 101, "stdout_tail": "", "stderr_tail": "tests::a ... FAILED",
-                "stdout_truncated": false, "stderr_truncated": false,
-                "failure_kind": "test_failure", "cwd": cwd, "command_summary": "cargo test",
-            }),
-        );
-        let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-        let validation = validation_summary_from_events(&summary.events, 20);
-        validation_delta_value(&validation)["comparison"]["scope_identity"]
-            .as_str()
-            .unwrap()
-            .to_string()
-    }
-    assert_ne!(
-        scope_with_cwd("crates/webcodex"),
-        scope_with_cwd("crates/webcodex-runner")
-    );
-}
-
-#[test]
-fn validation_delta_fail_a_to_fail_a_reports_still_failing() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "fail a to fail a");
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        test_output(0, 1, 0, &["tests::a"]),
-    );
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        test_output(0, 1, 0, &["tests::a"]),
-    );
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let validation = validation_summary_from_events(&summary.events, 20);
-    let delta = validation_delta_value(&validation);
-    let still = delta["failures"]["still_failing"].as_array().unwrap();
-    assert!(still.iter().any(|f| f["name"] == "tests::a"));
-    assert_eq!(
-        delta["failures"]["total_still_failing"], 1,
-        "count exceeds the bounded list total"
-    );
-}
-
-#[test]
-fn validation_delta_fail_a_to_fail_b_reports_newly_failed_and_resolved() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "fail a to fail b");
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        test_output(0, 1, 0, &["tests::a"]),
-    );
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        test_output(0, 1, 0, &["tests::b"]),
-    );
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let validation = validation_summary_from_events(&summary.events, 20);
-    let delta = validation_delta_value(&validation);
-    let newly = delta["failures"]["newly_failed"].as_array().unwrap();
-    let resolved = delta["failures"]["resolved"].as_array().unwrap();
-    assert!(newly.iter().any(|f| f["name"] == "tests::b"));
-    assert!(resolved.iter().any(|f| f["name"] == "tests::a"));
-    assert!(delta["failures"]["still_failing"]
-        .as_array()
-        .unwrap()
-        .is_empty());
-}
-
-#[test]
-fn validation_delta_multi_failure_partial_resolution() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "partial resolve");
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        test_output(0, 2, 0, &["tests::a", "tests::b"]),
-    );
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        test_output(0, 1, 0, &["tests::b"]),
-    );
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let validation = validation_summary_from_events(&summary.events, 20);
-    let delta = validation_delta_value(&validation);
-    assert!(delta["failures"]["resolved"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .any(|f| f["name"] == "tests::a"));
-    assert!(delta["failures"]["still_failing"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .any(|f| f["name"] == "tests::b"));
-    assert_eq!(delta["failures"]["total_still_failing"], 1);
-}
-
-// =========================================================================
-// Incomparable validation
-// =========================================================================
-
-#[test]
-fn validation_delta_unavailable_when_command_filter_changes() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "filter change");
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        json!({
-            "exit_code": 101,
-            "stdout_tail": "",
-            "stderr_tail": "tests::a ... FAILED",
-            "stdout_truncated": false,
-            "stderr_truncated": false,
-            "failure_kind": "validation_failed",
-            "cwd": "crates/webcodex",
-            "command_summary": "cargo test --lib tests::a",
-        }),
-    );
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        json!({
-            "exit_code": 101,
-            "stdout_tail": "",
-            "stderr_tail": "tests::b ... FAILED",
-            "stdout_truncated": false,
-            "stderr_truncated": false,
-            "failure_kind": "validation_failed",
-            "cwd": "crates/webcodex",
-            "command_summary": "cargo test --lib tests::b",
-        }),
-    );
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let validation = validation_summary_from_events(&summary.events, 20);
-    let delta = validation_delta_value(&validation);
-    assert_eq!(delta["comparison"]["status"], "unavailable");
-    assert_eq!(
-        delta["comparison"]["reason_code"],
-        "validation_scope_changed"
-    );
-}
-
-#[test]
-fn validation_delta_unavailable_when_cwd_changes() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "cwd change");
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        json!({
-            "exit_code": 101,
-            "stdout_tail": "",
-            "stderr_tail": "tests::a ... FAILED",
-            "stdout_truncated": false,
-            "stderr_truncated": false,
-            "failure_kind": "validation_failed",
-            "cwd": "crates/webcodex",
-            "command_summary": "cargo test",
-        }),
-    );
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        json!({
-            "exit_code": 101,
-            "stdout_tail": "",
-            "stderr_tail": "tests::b ... FAILED",
-            "stdout_truncated": false,
-            "stderr_truncated": false,
-            "failure_kind": "validation_failed",
-            "cwd": "crates/webcodex-runner",
-            "command_summary": "cargo test",
-        }),
-    );
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let validation = validation_summary_from_events(&summary.events, 20);
-    let delta = validation_delta_value(&validation);
-    assert_eq!(delta["comparison"]["status"], "unavailable");
-    assert_eq!(
-        delta["comparison"]["reason_code"],
-        "validation_scope_changed"
-    );
-}
-
-#[test]
-fn validation_delta_unavailable_with_no_previous_validation() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "single run");
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        true,
-        test_output(1, 0, 0, &[]),
-    );
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let validation = validation_summary_from_events(&summary.events, 20);
-    let delta = validation_delta_value(&validation);
-    assert_eq!(delta["comparison"]["status"], "unavailable");
-    assert_eq!(delta["comparison"]["reason_code"], "no_previous_validation");
-}
-
-#[test]
-fn validation_delta_unavailable_when_scope_identity_is_missing() {
-    // A check run that omits cwd/command_summary cannot be proven comparable.
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "missing scope");
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_check",
-        false,
-        json!({
-            "exit_code": 101,
-            "stdout_tail": "",
-            "stderr_tail": "error[E0308]: m",
-            "stdout_truncated": false,
-            "stderr_truncated": false,
-            "failure_kind": "validation_failed",
-        }),
-    );
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_check",
-        false,
-        json!({
-            "exit_code": 101,
-            "stdout_tail": "",
-            "stderr_tail": "error[E0277]: m",
-            "stdout_truncated": false,
-            "stderr_truncated": false,
-            "failure_kind": "validation_failed",
-        }),
-    );
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let validation = validation_summary_from_events(&summary.events, 20);
-    let delta = validation_delta_value(&validation);
-    assert_eq!(delta["comparison"]["status"], "unavailable");
-    assert_eq!(
-        delta["comparison"]["reason_code"],
-        "insufficient_scope_identity"
-    );
-}
-
-#[test]
-fn validation_delta_previous_evidence_incomplete_when_prior_run_unparsed() {
-    // The latest run is comparable in scope, but the prior run's output could
-    // not be parsed (no test result line, no failed-test line), so its failure
-    // identities are unknown and the delta is downgraded.
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "prior unparsed");
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        json!({
-            "exit_code": 101,
-            "stdout_tail": "",
-            "stderr_tail": "something unparseable happened",
-            "stdout_truncated": false,
-            "stderr_truncated": false,
-            "failure_kind": "test_failure",
-            "cwd": "crates/webcodex",
-            "command_summary": "cargo test --lib",
-        }),
-    );
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        true,
-        test_output(1, 0, 0, &[]),
-    );
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let validation = validation_summary_from_events(&summary.events, 20);
-    let delta = validation_delta_value(&validation);
-    assert_eq!(delta["comparison"]["status"], "unavailable");
-    assert_eq!(
-        delta["comparison"]["reason_code"],
-        "previous_evidence_incomplete"
-    );
-}
-
-#[test]
-fn validation_delta_zero_test_success_does_not_resolve_prior_test_failures() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "zero test resolve");
-    // Previous: one real test failure.
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        test_output(0, 1, 0, &["tests::a"]),
-    );
-    // Current: a "success" with zero tests run (no tests executed).
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        true,
-        json!({
-            "exit_code": 0,
-            "stdout_tail": "test result: ok. 0 passed; 0 failed; 0 ignored",
-            "stderr_tail": "",
-            "stdout_truncated": false,
-            "stderr_truncated": false,
-            "failure_kind": serde_json::Value::Null,
-            "cwd": "crates/webcodex",
-            "command_summary": "cargo test --lib",
-        }),
-    );
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let validation = validation_summary_from_events(&summary.events, 20);
-    let delta = validation_delta_value(&validation);
-    // The scope is comparable, so comparison status is available, but the
-    // failure identity diff must be unavailable (no false "resolved").
-    assert_eq!(delta["comparison"]["status"], "available");
-    assert_eq!(delta["failures"]["identity_status"], "unavailable");
-    assert_eq!(
-        delta["failures"]["identity_reason_code"],
-        "test_identity_unavailable"
-    );
-    assert!(
-        delta["failures"]["resolved"].as_array().unwrap().is_empty(),
-        "zero-test run must not falsely resolve prior failures"
-    );
-}
-
-// =========================================================================
-// Bounded output
-// =========================================================================
-
-#[test]
-fn changed_paths_are_deduped_bounded_and_report_total_and_truncated() {
-    // Real finished write events only. `closeout_work_projection` reads
-    // changed paths from `tool_call_finished` events, dedups them into a
-    // BTreeSet (deterministic sorted order), and `continuation_feedback`
-    // further truncates to MAX_CHANGED_PATHS (100) with truncated=true when
-    // the deduped total exceeds it. total_changed_paths is the real deduped
-    // count (kept <= 200 so the work projection's own cap does not distort it).
-    let runtime = test_runtime();
-
-    // Over the limit: 150 distinct paths + duplicates, all finished.
-    // Each write emits a started+finished pair (2 events), so batch 2 paths
-    // per write to keep the total event count inside the 200-event retained
-    // window; otherwise the oldest writes would be evicted and the deduped
-    // total would silently shrink. The ledger extraction (changed_paths_for_tool)
-    // reads `changes[].path` from each finished event — the same path the
-    // production closeout_work_projection reads.
-    let session = create_session(&runtime, "many paths");
-    add_instruction(&runtime, &session, "edit many", SessionMode::Normal);
-    let distinct: Vec<String> = (0..150).map(|i| format!("src/file_{i}.rs")).collect();
-    // 75 writes, each carrying 2 distinct paths; every 3rd write repeats an
-    // earlier path so the projection must collapse duplicates to one entry.
-    for i in 0..75 {
-        let p0 = distinct[i * 2].as_str();
-        let p1 = distinct[i * 2 + 1].as_str();
-        let batch: Vec<&str> = if i > 0 && i % 3 == 0 {
-            // repeat an earlier path alongside the two new ones
-            vec![p0, p1, distinct[(i - 1) * 2].as_str()]
-        } else {
-            vec![p0, p1]
-        };
-        record_write(&runtime, &session, &batch);
-    }
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    assert!(summary.events.len() <= 200);
-    let feedback = feedback_for(&runtime, &summary, "continued");
-    let changes = &feedback["attempt"]["changes"];
-    let total = changes["total_changed_paths"].as_u64().unwrap();
-    let truncated = changes["truncated"].as_bool().unwrap();
-    let returned: Vec<String> = changes["changed_paths"]
-        .as_array()
-        .unwrap()
-        .iter()
-        .map(|v| v.as_str().unwrap().to_string())
-        .collect();
-    assert_eq!(total, 150, "total is the real deduped count");
-    assert!(truncated, "over the limit => truncated=true");
-    assert_eq!(returned.len(), 100, "capped at MAX_CHANGED_PATHS");
-    // Deterministic sorted order (BTreeSet): returned prefix is sorted.
-    let mut sorted_distinct = distinct.clone();
-    sorted_distinct.sort();
-    assert_eq!(returned, sorted_distinct[..100].to_vec());
-    // No duplicate appears twice in the returned list.
-    let mut seen = std::collections::HashSet::new();
-    for path in &returned {
-        assert!(
-            seen.insert(path.clone()),
-            "duplicate path in output: {path}"
-        );
-    }
-
-    // Under the limit: nothing truncated, all returned.
-    let session_under = create_session(&runtime, "under limit");
-    add_instruction(&runtime, &session_under, "few edits", SessionMode::Normal);
-    for path in &["src/a.rs", "src/b.rs", "src/a.rs"] {
-        record_write(&runtime, &session_under, std::slice::from_ref(path));
-    }
-    let summary = runtime.sessions.summary(&session_under, Some(200)).unwrap();
-    let feedback = feedback_for(&runtime, &summary, "continued");
-    let changes = &feedback["attempt"]["changes"];
-    assert_eq!(changes["total_changed_paths"].as_u64().unwrap(), 2);
-    assert!(!changes["truncated"].as_bool().unwrap());
-    assert_eq!(changes["changed_paths"].as_array().unwrap().len(), 2);
-
-    // Exactly at the limit: not truncated, len == MAX. 50 writes x 2 paths
-    // = 100 distinct paths, ~102 events, well inside the window.
-    let session_at = create_session(&runtime, "at limit");
-    add_instruction(&runtime, &session_at, "exactly limit", SessionMode::Normal);
-    for i in 0..50 {
-        let p0 = format!("src/exact_{i}a.rs");
-        let p1 = format!("src/exact_{i}b.rs");
-        record_write(&runtime, &session_at, &[p0.as_str(), p1.as_str()]);
-    }
-    let summary = runtime.sessions.summary(&session_at, Some(200)).unwrap();
-    let feedback = feedback_for(&runtime, &summary, "continued");
-    let changes = &feedback["attempt"]["changes"];
-    assert_eq!(changes["total_changed_paths"].as_u64().unwrap(), 100);
-    assert!(
-        !changes["truncated"].as_bool().unwrap(),
-        "exactly at the limit is not truncated"
-    );
-    assert_eq!(changes["changed_paths"].as_array().unwrap().len(), 100);
-}
-
-#[test]
-fn continuation_feedback_output_is_bounded_in_size() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "bounded size");
-    add_instruction(&runtime, &session, "bounded", SessionMode::Normal);
-    for _ in 0..40 {
-        record_validation_event(
-            &runtime,
-            &session,
-            "cargo_test",
-            false,
-            test_output(0, 1, 0, &["tests::x"]),
-        );
-    }
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let feedback = feedback_for(&runtime, &summary, "continued");
-    let serialized = serde_json::to_string(&feedback).unwrap();
-    // The projection stays bounded even with many validation events.
-    assert!(
-        serialized.len() < 32_000,
-        "continuation feedback must stay bounded: {} bytes",
-        serialized.len()
-    );
-    // suggested_next_actions is bounded.
-    assert!(
-        feedback["attempt"]["suggested_next_actions"]
-            .as_array()
-            .map(|a| a.len() <= 8)
-            .unwrap_or(true),
-        "suggested_next_actions bounded"
-    );
-}
-
-// =========================================================================
-// Message board
-// =========================================================================
-
-#[test]
-fn guidance_counts_open_messages_without_changing_status() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "guidance board");
-    add_instruction(&runtime, &session, "with guidance", SessionMode::Normal);
-    post_session_message(&runtime, &session, "guidance", "review the new boundary");
-    post_session_message(&runtime, &session, "risk", "risk A");
-    post_session_message(&runtime, &session, "todo", "todo A");
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let discussion = runtime
-        .sessions
-        .discussion_summary(&session, Some(20))
-        .unwrap();
-    let feedback = feedback_for_with_discussion(&runtime, &summary, &discussion, "continued");
-    assert_eq!(feedback["attempt"]["guidance"]["open_count"], 1);
-    assert_eq!(feedback["attempt"]["guidance"]["open_risk_count"], 1);
-    assert_eq!(feedback["attempt"]["guidance"]["open_todo_count"], 1);
-    assert_eq!(feedback["attempt"]["guidance"]["latest_open_kind"], "todo");
-
-    // Reading the projection must not resolve or mark-read anything.
-    let after = runtime
-        .sessions
-        .discussion_summary(&session, Some(20))
-        .unwrap();
-    assert_eq!(after.counts.guidance, 1);
-    assert_eq!(after.counts.risk, 1);
-    assert_eq!(after.counts.todo, 1);
-}
-
-#[test]
-fn resolved_guidance_is_not_counted_as_open() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "resolved guidance");
-    add_instruction(&runtime, &session, "resolve", SessionMode::Normal);
-    post_session_message(&runtime, &session, "guidance", "resolved guidance");
-    // Resolve the only guidance message.
-    let listed = runtime
-        .sessions
-        .list_messages(
-            &session,
-            sessions::ListSessionMessagesFilter {
-                kind: Some(sessions::SessionMessageKind::Guidance),
-                status: None,
-                message_id: None,
-                reply_to: None,
-                limit: None,
-            },
-        )
-        .unwrap();
-    let message_id = listed.first().unwrap().message_id.clone();
-    runtime
-        .sessions
-        .resolve_message(&session, &message_id, Some("done".to_string()))
-        .unwrap();
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let discussion = runtime
-        .sessions
-        .discussion_summary(&session, Some(20))
-        .unwrap();
-    let feedback = feedback_for_with_discussion(&runtime, &summary, &discussion, "continued");
-    assert_eq!(feedback["attempt"]["guidance"]["open_count"], 0);
-}
-
-// =========================================================================
-// Jobs
-// =========================================================================
-
-#[test]
-fn jobs_block_reports_recovering_not_healthy_running() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "jobs recovering");
-    add_instruction(&runtime, &session, "jobs", SessionMode::Normal);
-    let jobs = json!({
-        "active_count": 1,
-        "running_count": 1,
-        "recovering_count": 1,
-        "terminal_pending_count": 0,
-        "truncated": false,
-        "recent": [{"job_id": "j1", "status": "recovering"}],
-    });
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let feedback = feedback_for_with_jobs(&runtime, &summary, &jobs, "continued");
-    assert_eq!(feedback["attempt"]["jobs"]["recovering_count"], 1);
-    assert_eq!(feedback["attempt"]["jobs"]["recovery_state"], "recovering");
-    // recovering must not masquerade as healthy running.
-    assert_ne!(feedback["attempt"]["jobs"]["recovery_state"], "healthy");
-    assert_ne!(feedback["attempt"]["jobs"]["recovery_state"], "active");
-}
-
-#[test]
-fn jobs_block_terminal_pending_does_not_become_healthy_running() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "jobs terminal pending");
-    add_instruction(&runtime, &session, "jobs", SessionMode::Normal);
-    let jobs = json!({
-        "active_count": 1,
-        "running_count": 0,
-        "recovering_count": 0,
-        "terminal_pending_count": 1,
-        "truncated": false,
-        "recent": [{"job_id": "j1", "status": "stop_requested"}],
-    });
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let feedback = feedback_for_with_jobs(&runtime, &summary, &jobs, "continued");
-    assert_eq!(feedback["attempt"]["jobs"]["terminal_pending_count"], 1);
-    assert_eq!(
-        feedback["attempt"]["jobs"]["recovery_state"],
-        "terminal_pending"
-    );
-}
-
-#[test]
-fn jobs_block_none_when_no_active_jobs() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "no jobs");
-    add_instruction(&runtime, &session, "jobs", SessionMode::Normal);
-    let jobs = json!({
-        "active_count": 0,
-        "running_count": 0,
-        "recovering_count": 0,
-        "terminal_pending_count": 0,
-        "truncated": false,
-        "recent": [],
-    });
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let feedback = feedback_for_with_jobs(&runtime, &summary, &jobs, "continued");
-    assert_eq!(feedback["attempt"]["jobs"]["recovery_state"], "none");
-    assert_eq!(
-        feedback["attempt"]["jobs"]["latest_job_status"],
-        "not_observed"
-    );
-}
-
-#[test]
-fn jobs_block_recovering_hidden_beyond_truncated_recent_still_reported() {
-    // Regression: counts come from the aggregate, not the truncated `recent`
-    // list. A recovering job hidden beyond the recent window must still drive
-    // recovery_state = recovering (never masquerade as healthy/active).
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "jobs truncated recent");
-    add_instruction(&runtime, &session, "jobs", SessionMode::Normal);
-    let jobs = json!({
-        "active_count": 5,
-        "running_count": 5,
-        "recovering_count": 1,
-        "terminal_pending_count": 0,
-        "truncated": true,
-        "recent": [{"job_id": "j1", "status": "running"}, {"job_id": "j2", "status": "running"}],
-    });
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let feedback = feedback_for_with_jobs(&runtime, &summary, &jobs, "continued");
-    assert_eq!(feedback["attempt"]["jobs"]["recovering_count"], 1);
-    assert_eq!(feedback["attempt"]["jobs"]["recent_truncated"], true);
-    assert_eq!(feedback["attempt"]["jobs"]["recovery_state"], "recovering");
-}
-
-// =========================================================================
-// Canonical coding workflow integration: pre-instruction attempt
-// =========================================================================
-
-use crate::tool_runtime::{StartupDetail, ToolCall, ToolResult};
 
 async fn diagnostic_coding_workflow(
     runtime: &ToolRuntime,
@@ -2019,67 +697,6 @@ async fn handoff_guidance_read_does_not_resolve_messages() {
     assert_eq!(after.counts.guidance, 1);
     assert_eq!(after.counts.todo, 1);
 }
-
-#[test]
-fn continuation_feedback_does_not_leak_raw_output_or_commands() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "no leak");
-    add_instruction(&runtime, &session, "secret command", SessionMode::Normal);
-    record_validation_event(
-        &runtime,
-        &session,
-        "cargo_test",
-        false,
-        json!({
-            "exit_code": 101,
-            "stdout_tail": "SENSITIVE_TOKEN=abc123",
-            "stderr_tail": "tests::a ... FAILED",
-            "stdout_truncated": false,
-            "stderr_truncated": false,
-            "failure_kind": "validation_failed",
-            "cwd": "crates/webcodex",
-            "command_summary": "cargo test --lib",
-        }),
-    );
-
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let feedback = feedback_for(&runtime, &summary, "continued");
-    let serialized = serde_json::to_string(&feedback).unwrap();
-    assert!(
-        !serialized.contains("SENSITIVE_TOKEN"),
-        "continuation feedback leaked a sensitive token"
-    );
-    assert!(
-        !serialized.contains("stdout_tail") && !serialized.contains("stderr_tail"),
-        "continuation feedback leaked raw output field names"
-    );
-    assert!(
-        !serialized.contains("command_summary"),
-        "continuation feedback leaked command text"
-    );
-}
-
-#[test]
-fn fresh_session_reports_not_applicable() {
-    let runtime = test_runtime();
-    let session = create_session(&runtime, "fresh");
-    let summary = runtime.sessions.summary(&session, Some(200)).unwrap();
-    let feedback = feedback_for(&runtime, &summary, "created");
-    assert_eq!(feedback["status"], "not_applicable");
-    assert_eq!(feedback["reason_code"], "fresh_session");
-    assert_eq!(
-        feedback["attempt"]["exploration"]["observed_paths"],
-        json!([])
-    );
-    assert_eq!(
-        feedback["attempt"]["exploration"]["latest_tool"],
-        Value::Null
-    );
-}
-
-// =========================================================================
-// Real entry integration: validation_summary
-// =========================================================================
 
 #[tokio::test]
 async fn validation_summary_surfaces_validation_delta_without_shell_or_new_events() {
@@ -2474,30 +1091,6 @@ fn record_exploration_event(
         (!success).then_some("exploration failed"),
         None,
     );
-}
-
-fn locations_output(path: &str, locations: &[&str]) -> Value {
-    json!({
-        "project": "test-project",
-        "path": path,
-        "query_position": {"line": 3, "column": 5},
-        "locations": locations
-            .iter()
-            .enumerate()
-            .map(|(index, path)| json!({
-                "path": path,
-                "range": {
-                    "start": {"line": index + 1, "column": 1},
-                    "end": {"line": index + 1, "column": 2}
-                }
-            }))
-            .collect::<Vec<_>>(),
-        "total_results": locations.len(),
-        "returned_count": locations.len(),
-        "truncated": false,
-        "external_results_omitted": 0,
-        "invalid_results_omitted": 0
-    })
 }
 
 /// Project-scoped variants of the helpers above for real entry tests that
